### 2020-09-10 #2658 @whedon
[PRE REVIEW]: Minimalist And Customizable Optimization Package
**Submitting author:** @jbuisine (<a href="http://orcid.org/0000-0001-6071-744X">Jérôme BUISINE</a>)
**Repository:** <a href="https://github.com/jbuisine/macop" target ="_blank">https://github.com/jbuisine/macop</a>
**Version:** v1.0.0
