### 2020-09-10 #2658 @danielskatz
👋 @jbuisine - thanks for your submission.  The next step will be an internal discussion by JOSS editors about the scope, which will take about a week, then we will get back to you about the next steps
