### 2020-10-02 #2658 @Kevin-Mattheus-Moerman
@jbuisine upon initial inspection our board has decided this work may be in scope for JOSS. I will no proceed find an editor to handle this work. We do feel the paper needs a lot of improvement so please do work with the editor to get it improved, likely prior to the formal review stage.
