### 2020-10-05 #2658 @melissawm
Hello, @jbuisine ! Before we proceed into the review stage, I'd like to suggest a careful review of JOSS's [Submission Requirements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). More specifically, I would suggest a stronger statement of need and a description of where this software fits into the current landscape, as it's not clear what differentiates it from existing optimization/OR frameworks. Also, if this was created to solve a particular research problem, I would encourage it to be cited somewhere in the paper. Please let me know if you have any questions or comments!
