### 2020-10-06 #2658 @jbuisine
Hello @melissawm, thank you for your advice, I will take it into account and modify the document in order to situate the package in the bibliography!
