+++ paper.md
@@ -3,6 +3,7 @@ title: 'Minimalist And Customizable Optimization Package'
 tags:
   - Python
   - Operations Research
+  - Mono-objective
   - Multi-objective
 authors:
   - name: Jérôme BUISINE
@@ -11,7 +12,7 @@ authors:
 affiliations:
  - name: Univ. Littoral Côte d’Opale, LISIC Calais, France, F-62100
    index: 1
-date: 9 September 2020
+date: 11 October 2020
 bibliography: paper.bib
 
 # Optional fields if submitting to a AAS journal too, see this blog post:
@@ -24,19 +25,19 @@ bibliography: paper.bib
 
 `Macop` for `Minimalist And Customizable Optimization Package` is an optimization Python package which not implement the whole available algorithms in the literature but let you the possibility to quickly develop and test your own algorithm and strategies. The main objective of this package is to be the most flexible as possible and hence, to offer a maximum of implementation possibilities.
 
-An underlying objective is to enable this package to be used in educational contexts as well. Allowing students to quickly develop their own algorithms.
+An underlying objective is to enable this package to be used for quick implementations and in educational contexts as well. Allowing students to quickly develop their own algorithms.
 
 # Motivation
 
-During thesis work, the search for a solution with complex evaluation was necessary. The assessment in question consisted of evaluating a model fited with selected subset of available features from a features set. The solution was therefore the new model obtained and its fitness, the score obtained on this test basis.
+Most of the operational research libraries developed in Python offer users either problem and algorithm implementations where it is possible to choose parameters to obtain optimal (or near optimal) results such as [pyomo](http://www.pyomo.org/documentation), [PuLP](https://pypi.org/project/PuLP/), [pyOpt](http://www.pyopt.org/), [PySCIPOpt](https://github.com/SCIP-Interfaces/PySCIPOpt) or, libraries targeted to a specific problem or algorithm such as [simanneal](https://github.com/perrygeo/simanneal).
 
-Exploring all the solutions was not feasible given the large amount of exploration space available. Otherwise it would have been preferable to use a Recursive Features Elimination method [@DBLP:journals/remotesensing/PullanagariKY18; @DBLP:conf/icmla/ChenJ07].
+During thesis work, the search for a binary solution with complex evaluation was necessary. The assessment in question consisted of evaluating a model fited with selected subset of available features from a features set. Binary solution was used as represation for selected or non-selected feature from available set of features. The solution was therefore a new model obtained and its fitness which is the score obtained on test data set. Exploring all the solutions was not feasible given the large amount of exploration space available. Otherwise it would have been preferable to use a Recursive Features Elimination method [@DBLP:journals/remotesensing/PullanagariKY18; @DBLP:conf/icmla/ChenJ07].
 
-This is why it was preferred to focus on operational research and to develop this python package. Even if the optimal solution is not found, a good solution is still sufficient.
+This is why it was preferred to focus on operational research methods, even if the optimal solution is not found, a good solution is still sufficient for the study case. Libraries available in the literature did not allow this kind of implementation of an evaluation function quickly and this is why the development of a Python package has been undertaken. This package was then adapted to correspond to a formalism of implementing algorithms and problems from a simple structure, hence the term minimalist for Macop.
 
 # Description
 
-At the beginning of the development of this tool, the idea of making it as modular as possible was topical.
+At the beginning of the development of this tool, the idea of making it as modular as possible was topical. By dividing the library into sub-module forms considered to be the most important to build and solve an optimisation problem.
 
 The package consists of several modules:
 
@@ -46,10 +47,17 @@ The package consists of several modules:
[…]
