+++ examples/knapsackMultiExample.py
@@ -80,13 +80,13 @@ def main():
     policy = UCBPolicy(operators, _C=100, _exp_rate=0.2)
 
     # pass list of evaluators
-    algo = MOEAD(20, 5, init, [evaluator1, evaluator2, evaluator2, evaluator2], operators, policy, validator, _maximise=True)
+    algo = MOEAD(20, 5, init, [evaluator1, evaluator2], operators, policy, validator, _maximise=True)
     
     algo.addCallback(MultiCheckpoint(_every=5, _filepath=mo_checkpoint_path))
     algo.addCallback(ParetoCheckpoint(_every=5, _filepath=pf_checkpoint_path))
     algo.addCallback(UCBCheckpoint(_every=5, _filepath=ucb_checkpoint_path))
 
-    paretoFront = algo.run(10000)
+    paretoFront = algo.run(1000)
 
     print("Pareto front is composed of", len(paretoFront), "solutions")
 
