### 2020-10-11 389c212
add test when building package
