+++ build.sh
@@ -7,6 +7,7 @@ yapf -ir -vv macop
 # Build rawls package
 echo "Build package..."
 python setup.py build
+python setup.py test
 
 echo "Build documentation..."
 rm -r docs/source/macop
