### 2020-10-11 847a1b6
JOSS paper updated
