### 2020-10-11 #2658 @jbuisine
Hello @melissawm, thanks to your advices, I generated a new version of the submited paper. I detailed the purpose and the need of this developed package during my thesis.

Do not hesitate to tell me if something is not very clear or if something seems to be missing.

Thanks in advance.
