### 2020-10-22 6f595e4
Change sentences structures
