+++ paper.md
@@ -29,7 +29,7 @@ An underlying objective is to enable this package to be used for quick implement
 
 # Motivation
 
-Most of the operational research libraries developed in Python offer users either problem and algorithm implementations where it is possible to choose parameters to obtain optimal (or near optimal) results such as [pyomo](http://www.pyomo.org/documentation), [PuLP](https://pypi.org/project/PuLP/), [pyOpt](http://www.pyopt.org/), [PySCIPOpt](https://github.com/SCIP-Interfaces/PySCIPOpt), or, libraries targeted to a specific problem or algorithm such as [simanneal](https://github.com/perrygeo/simanneal).
+Most of the operational research libraries developed in Python offer users either problems and algorithms where it is possible to choose parameters to obtain optimal (or near optimal) results such as [pyomo](http://www.pyomo.org/documentation), [PuLP](https://pypi.org/project/PuLP/), [pyOpt](http://www.pyopt.org/), [PySCIPOpt](https://github.com/SCIP-Interfaces/PySCIPOpt), or, libraries targeted to a specific problem or algorithm such as [simanneal](https://github.com/perrygeo/simanneal).
 
 During thesis work, the search for a solution with complex evaluation was necessary. The assessment in question consisted of evaluating a model fited with selected subset of available features from a features set. Binary solution was used as represation for selected or non-selected feature from available set of features. The solution was therefore a new model obtained and its fitness which is the score obtained on test data set. Exploring all the solutions was not feasible given the large amount of exploration space available. Otherwise it would have been preferable to use a Recursive Features Elimination (RFE) method [@DBLP:journals/remotesensing/PullanagariKY18; @DBLP:conf/icmla/ChenJ07].
 
@@ -45,7 +45,7 @@ The package consists of main several modules:
 - `evaluator`: contains all the implemented evaluation functions.
 - `solutions`: all declared solutions classes used to represent problem data.
 - `operators`: mutators, crossovers update of solution. This module also has policies classes to manage the way of update and use solution.
-- `callbacks`: contains Callback class implementations for making callback instructions every number of evaluations.
+- `callbacks`: contains Callback classes for making callback instructions every number of evaluations.
 
 The primary advantage of using Python is that it allows you to dynamically add new members within the new implemented solution or algorithm classes. This of course does not close the possibilities of extension and storage of information within solutions and algorithms. It all depends on the need in question.
 
@@ -57,7 +57,7 @@ A hieararchy between dependent algorithms is also available, based on a parent/c
 
 The mono-objective Iterated Local Search [@DBLP:books/sp/03/LourencoMS03] algorithm which aims to perform local searches (child algorithms linked to the main algorithm) and then to explore again (explorations vs. exploitation trade-off). On the multi-objective side, the MOEA/D algorithm [@DBLP:journals/tec/ZhangL07] has been implemented too using the weighted-sum of objectives method (Tchebycheff approach can also be used [@DBLP:journals/cor/AlvesA07]). This algorithm aims at decomposing the multi-objective problem into $mu$ single-objective problems in order to obtain the pareto front [@kim2005adaptive] where single-objective problems are so-called child algorithms linked to the multi-objective algorithm.
 
-The main purpose of these algorithm implementations is to show the possibilities of operational search algorithm implementations based on the minimalist structure of the library.
+The main purpose of these developed algorithms is to show the possibilities of operational search algorithm implementations based on the minimalist structure of the library.
 
 ## Available solutions
 
@@ -73,11 +73,11 @@ The use of callback instance allows both to do an action every $k$ evaluations o
 
 ## Documentation
 
-Documentation with examples for mono and multi-objective implementation is available at [https://jbuisine.github.io/macop/](https://jbuisine.github.io/macop/).
[…]
