### 2020-10-22 #2658 @melissawm
Hello, @jbuisine ! Thank you for the modifications. I noticed a number of typos, which I think might be corrected before the review stage - maybe going through an autocorrect feature. Another issue that I think might be worth noting is that the problem (or problems) being solved by the software are not explicitly mentioned in the software paper. From the context, I understand you are doing global optimization in an OR setting - is that correct? I think that is worth mentioning somewhere.

Finally: do you have any suggestions for potential reviewers from the list mentioned in the first comment on this issue? Thank you!
