### 2020-10-23 c911dc6
Add introduction to discrete and continuous problem
