### 2020-10-23 19dc4cf
fix LaTeX math command
