### 2020-10-23 287df28
update math syntax on paper
