+++ setup.py
@@ -81,7 +81,7 @@ class TestCommand(distutils.command.check.check):
 setup(
     name='macop',
     version='1.0.0',
-    description='Minimalist And Customizable Optimization Package',
+    description='Minimalist And Customisable Optimisation Package',
     long_description=open('README.md').read(),
     long_description_content_type='text/markdown',
     classifiers=[
