### 2020-10-23 #2658 @jbuisine
Hello @melissawm,

Thank you again for your feedback. I have updated the paper accordingly, I hope this is more understandable.

That's right, the idea is to propose a generic tool allowing to refrain from some recurrent concepts when implementing operational search algorithms (solution evaluation, hierarchical management of algorithms, backups, algorithm reloading). As the tool is not specific, the user could implement his algorithm without having to overload or be limited when faced with the specificity of an implementation. It is a tool developed within the framework of the thesis which allowed me to implement some specific algorithms quickly and which I would like to share.

Concerning the suggestions of reviewers, I was thinking about : `stsievert`, `jakobbossek` and `rowanc1` (I do not quote them directly here to avoid unwanted spam).

Thanks.
