### 2020-10-30 #2658 @melissawm
Hello @torressa, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html Let me know if you need any more information. Thank you!
