### 2020-11-02 #2812 @whedon
[REVIEW]: Minimalist And Customizable Optimization Package
**Submitting author:** @jbuisine (<a href="http://orcid.org/0000-0001-6071-744X">Jérôme BUISINE</a>)
**Repository:** <a href="https://github.com/jbuisine/macop" target ="_blank">https://github.com/jbuisine/macop</a>
**Version:** v1.2.0

    * ~~Specifically, this claim is made in paper.md: "Solutions modeling continuous problems can also be created by the anyone who wants to model his own problem", but I'm not seeing a continuous optimization problem in the examples or tests.~~
