### 2020-11-16 #2812 @stsievert
@melissawm I have a review waiting. It looks like I didn't accept the invitation to collaborate on this repo in time, so I can't edit https://github.com/openjournals/joss-reviews/issues/2812#issue-734575770. Could you resend that invitation?
