### 2020-11-19 #2812 @stsievert
My initial review: I do not believe this software project meets the [JOSS review requirements]. I welcome questions and comments from @jbuisine so provide clarification or be corrected on points below.

The most significant comments/questions are below. I have also edited https://github.com/openjournals/joss-reviews/issues/2812#issue-734575770 to check various boxes* and add some notes on specific items.

* **The tests are not sufficient.** I only found [a single test][test] in setup.py, and not a single doctest [as claimed]. In this one test, there are many untested classes.
* **Target audience and statement of need.** [The main documentation][docs] do not identify a target audience or a statement of need. Only one feature of the software, maximum flexibility, is described. Who would want that and why? In paper.md, I see that there's a need for maximum flexibility mentioned in the paper, but why would a user want if they're not doing "thesis work"?

The documentation has some specific problems: it's very general and not detailed, is missing a good API description.

* **Generality:** optimization is very general. When is Macop not appropriate?
    * For example, could I use it for my favorite deep learning problem? Could it be used for a continuous convex optimization?
    * Or what if the optimization problem I'm using requires sparse arrays or GPUs?
* **Detailed:** the only information core classes have is their name – past that, there's not much detail.
    * For example, the documentation for `IteratedLocalSearch` is basically "[this class is] used to avoid local optima and increave EvE (Exploration vs Exploitation) compromise".  What's "EvE", what effect does it have, and how do the various parameters balance it?
    * For example, what do `C` and `exp_rate` do in [UCBPolicy]? Does increasing `C` increase or decrease the importance of exploration? How does changing the value of `exp_rate` change the exploration/exploitation tradeoff?
* **API**: it's not clear what different policies, mutators and crossovers are used for. What's the data flow of one example? Detailing the input/output of these modules would be useful. I'd expect to see a clear definitions for non-experts.

<summary>Some specific questions/suggestions/typos for paper.md:</summary>

* "generic and implemented OR algorithms" This is the first definition of "OR".
* "Tools for modelling and solving discrete and continuous problems are proposed in the literature." This needs at least one citation.
* "Allowing students to quickly develop their own algorithms." This sentence is a sentence fragment: it doesn't express a complete thought.
* "finding a point $x \in X$ en that has" What does "en" mean?
* "implement the whole available algorithms in the literature" -> "which doesn't implement every algorithm in the literature"
* "but let you the possibility to quickly develop and test" -> "but provides the ability to ..."
* "The main objective of this package is to be the most flexible as possible and hence, to offer a maximum of implementation possibilities." This sentence doesn't make sense. I think you're trying to convey "The main objective of this package is to provide maximum flexibility, which allows for easy experimentation in implementation."
* The second paragraph in motivation needs cleaning up -- I'm having a hard time deciphering "Binary solution was used as appreciation for selected or non-selected feature from the available set of features. The solution was therefore a new model obtained and its fitness which is the score obtained on the test data set. "
* $\mathbb{N}$ should be used instead of $\mathbf{N}$

<sup>* a checked box ([x]) means "satisfied", an unchecked box ([ ]) means "not reviewed" and a crossed box (~~[ ]~~) means "not satisfied". A partially satisfied requirement is indicated with "[.]"</sup>

[UCBPolicy]:https://jbuisine.github.io/macop/build/html/macop/macop.operators.policies.UCBPolicy.html
[docs]:https://jbuisine.github.io
[as claimed]:https://github.com/jbuisine/macop/blob/287df287e23c7e4f07e90dfcc0a99ef247f5c6b5/CONTRIBUTING.md#test-implementation
[test]:https://github.com/jbuisine/macop/blob/287df287e23c7e4f07e90dfcc0a99ef247f5c6b5/setup.py#L7-L78
[JOSS review requirements]:https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
