### 2020-11-22 #2812 @torressa
Hi @jbuisine thanks for the submission!
Here's my review, I welcome responses and questions in any of the following points. Also, I may have more comments in the future.

### Major Issues (general)

- Documentation: bad.
  - The api is present but badly documented (e.g. "Iterated Local Search Algorithm implementation") and not up to date with the master branch. It needs more elaborate explanations of the implementations and citations where appropriate.
  - The README and paper do not present an accurate description of what Macop actually is. For this you have to go through the code. Clearer technical description (in paper too)
  - The uses of the toolbox has to be better defined, where is it better to use Macop over exact or other methods.
- Testing: bad. Unit tests missing (there is one test in the setup.py file)
- LICENSE not completed.
- Possibly not a substantial scholarly effort
  - I've found the bulk of the algorithmic implementation to be mostly contained in July 2019 and August-October 2020. The rest of the work is re-structuring and packaging. Therefore, raising the question of whether the time spent on this project was [greater than 3 months](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort).
  - To add to this, as reported by [whedon](https://github.com/openjournals/joss-reviews/issues/2658#issuecomment-690245174) there's 1114 lines of python code. Given the structure of the code (with the nested inheritance and redefinition of functions), it is hard to say whether this ticks the "Substantial scholarly effort" box.

- Software paper:
  - State of the field: there is no mention of other similar software packages. The author comments on well-known general optimisation packages (which need a proper citation), but nothing related to the methods used in Macop. In a quick google-search I found a more closely related repo: [Solid](https://github.com/100/Solid) which  provides multiple derivate-free algorithms with user-defined objectives. The author probably knows more about this but I suspect there must be others, either way, it should be stated in the paper.
  - Quality of writing: the paper needs rereading to correct minor grammatical / orthographical errors. I will provide a list once the other comments have been addressed.

### Major Issues (code):
- Software usage: the usage of library is not very clear. How does one extract the final solution without looking at the logs? Not the user defined criteria like the score/fitness, but the final solution (i.e. in the knapsack example, the items in the knapsack).
- Why is abstract class inheritance only sometimes used? This seems like the right type of inheritance for all the classes.
- `macop/evaluators`: contains a single example and not an implementation
- `macop/operators`: triple nested inheritance makes it hard to determine where functions are actually being implemented (e.g. `UCBPolicy` inherits from `Policy` which inherits from `Operator`). For such simple functionality, e.g definining the method `apply`, this seems a bit overkill. Additionally, inheritance is not used e.g. `macop/operators/policies/UCBPolicy.py` could be a standalone class.
- `macop/solutions/`: Same as above, inheritance doesn't do much as functions are overloaded and arguments not passed to the parent. The purpose of these functions is really not clear to me.
- In general, the module like structure is really confusing, specially given the small and simple implementations. This makes the imports list to extend a fair bit. I understand that this structure may have been laid out for additional functionality to be introduced. However, if no more work is to be done, to help people understand the code and encourage contributions I suggest a major rework of the code to simplify the structure.

### Minor comments
- Please get rid of the build files for the documentation. readthedocs can build the docs on commits to master if you set it up.
- Relative imports are [discouraged by PEP 328](https://www.python.org/dev/peps/pep-0328/#id9) *"With the shift to absolute imports, the question arose whether relative imports should be allowed at all"*.
- Remove leading `_` for all input parameters, this is common practice for private parameters.
- `macop/callbacks`: they are technically a callback, but it seems more like a custom logger
- Is there any way to disable the progress output being printed to screen? I can imagine this needs disabling in normal applications
- The knapsack is a nice an simple example, however, I don't think it showcases the best use of this library. As it takes longer to arrive at the optimal solution with 30 elements with Macop than traditional exact methods.
- In a [standard](https://packaging.python.org/tutorials/packaging-projects/) python package, the init file in the main directory is not needed. init file in `macop/` should contain module imports
- Contributions file is really long while explanations on how to contribute are short.
- Examples [docs page](https://jbuisine.github.io/macop/build/html/examples.html#) is really long (half of it is imports and multi-line comments)
- In README:
  - Tags in readme should link to ci results and external urls, not images.
  - "OR algorithms"? list them.
