### 2020-11-24 #2812 @jbuisine
Thanks to @stsievert and @torressa for your remarks. @melissawm Yes it is something I can work on. I will see to take into consideration each of the remarks and update the Python package accordingly (documentation, examples, structural problems...). I will also take the time to answer each of your questions and queries about the package during the week.
