### 2020-12-02 832d00f
add of some new local search algorithms
