+++ macop/algorithms/mono/HillClimberFirstImprovment.py
@@ -1,4 +1,4 @@
-"""Local Search algorithm
+"""Hill Climber First Improvment algorithm starting from new solution and explore using neighborhood and loop over the best one obtained from neighborhood search space
 """
 
 # main imports
@@ -8,8 +8,12 @@ import logging
 from ..Algorithm import Algorithm
 
 
-class LocalSearch(Algorithm):
-    """Local Search used as exploitation optimisation algorithm
+class HillClimberFirstImprovment(Algorithm):
+    """Hill Climber First Improvment used as quick exploration optimisation algorithm
+
+    This algorithm do a neighborhood exploration of a new generated solution (by doing operation on the current solution obtained) in order to find a better solution from the neighborhood space.
+    Then replace the current solution by the first one from the neighbordhood space which is better than the current solution.
+    Do these steps until a number of evaluation (stopping criterion) is reached.
 
     Attributes:
         initalizer: {function} -- basic function strategy to initialize solution
@@ -22,27 +26,24 @@ class LocalSearch(Algorithm):
         bestSolution: {Solution} -- best solution found so far during running algorithm
         callbacks: {[Callback]} -- list of Callback class implementation to do some instructions every number of evaluations and `load` when initializing algorithm
     """
-    def run(self, _evaluations):
+    def run(self, evaluations):
         """
         Run the local search algorithm
 
         Args:
-            _evaluations: {int} -- number of Local search evaluations
+            evaluations: {int} -- number of Local search evaluations
             
         Returns:
             {Solution} -- best solution found
         """
 
         # by default use of mother method to initialize variables
-        super().run(_evaluations)
+        super().run(evaluations)
 
-        if self.parent:
-            self.bestSolution = self.parent.bestSolution
-
-        # initialize current solution
+        # initialize current solution and best solution
         self.initRun()
 
-        solutionSize = self.currentSolution.size
+        solutionSize = self._currentSolution.size
 
         # local search algorithm implementation
         while not self.stop():
@@ -50,24 +51,26 @@ class LocalSearch(Algorithm):
             for _ in range(solutionSize):
 
                 # update current solution using policy
-                newSolution = self.update(self.currentSolution)
+                newSolution = self.update(self._currentSolution)
 
-                # if better solution than currently, replace it
+                # if better solution than currently, replace it and stop current exploration (first improvment)
                 if self.isBetter(newSolution):
-                    self.bestSolution = newSolution
+                    self._bestSolution = newSolution
+                    break
 
                 # increase number of evaluations
                 self.increaseEvaluation()
 
                 self.progress()
-                logging.info("---- Current %s - SCORE %s" %
-                             (newSolution, newSolution.fitness()))
+                logging.info(f"---- Current {newSolution} - SCORE {newSolution.fitness()}")
 
                 # stop algorithm if necessary
                 if self.stop():
                     break
 
-        logging.info("End of %s, best solution found %s" %
-                     (type(self).__name__, self.bestSolution))
+            # set new current solution using best solution found in this neighbor search
+            self._currentSolution = self._bestSolution
 
-        return self.bestSolution
+        logging.info(f"End of {type(self).__name__}, best solution found {self._bestSolution}")
+
+        return self._bestSolution
