+++ macop/algorithms/Algorithm.py
@@ -12,6 +12,16 @@ from ..utils.color import macop_text, macop_line, macop_progress
 class Algorithm():
     """Algorithm class used as basic algorithm
 
+    This class enables to manage some common usages of operation research algorithms:
+    - initialization function of solution
+    - validator function to check if solution is valid or not (based on some criteria)
+    - evaluation function to give fitness score to a solution
+    - operators used in order to update solution during search process
+    - policy process applied when choosing next operator to apply
+    - callbacks function in order to do some relative stuff every number of evaluation or reload algorithm state
+    - parent algorithm associated to this new algorithm instance (hierarchy management)
+
+
     Attributes:
         initalizer: {function} -- basic function strategy to initialize solution
         evaluator: {function} -- basic function in order to obtained fitness (mono or multiple objectives)
@@ -19,45 +29,52 @@ class Algorithm():
         policy: {Policy} -- Policy class implementation strategy to select operators
         validator: {function} -- basic function to check if solution is valid or not under some constraints
         maximise: {bool} -- specify kind of optimisation problem 
-        currentSolution: {Solution} -- current solution managed for current evaluation
+        currentSolution: {Solution} -- current solution managed for current evaluation comparison
         bestSolution: {Solution} -- best solution found so far during running algorithm
         callbacks: {[Callback]} -- list of Callback class implementation to do some instructions every number of evaluations and `load` when initializing algorithm
         parent: {Algorithm} -- parent algorithm reference in case of inner Algorithm instance (optional)
     """
     def __init__(self,
-                 _initalizer,
-                 _evaluator,
-                 _operators,
-                 _policy,
-                 _validator,
-                 _maximise=True,
-                 _parent=None):
+                 initalizer,
+                 evaluator,
+                 operators,
+                 policy,
+                 validator,
+                 maximise=True,
+                 parent=None):
 
-        self.initializer = _initalizer
-        self.evaluator = _evaluator
-        self.operators = _operators
-        self.policy = _policy
-        self.validator = _validator
-        self.callbacks = []
-        self.bestSolution = None
-        self.currentSolution = None
+        # protected members intialization
+        self._initializer = initalizer
+        self._evaluator = evaluator
+        self._operators = operators
+        self._policy = policy
+        self._validator = validator
+        self._callbacks = []
+        self._bestSolution = None
+        self._currentSolution = None
 
         # by default
-        self.numberOfEvaluations = 0
-        self.maxEvaluations = 0
+        self._numberOfEvaluations = 0
+        self._maxEvaluations = 0
 
         # other parameters
-        self.parent = _parent  # parent algorithm if it's sub algorithm
+        self._parent = parent  # parent algorithm if it's sub algorithm
 
         #self.maxEvaluations = 0 # by default
-        self.maximise = _maximise
+        self._maximise = maximise
 
-        # track reference of algo into operator (keep an eye into best solution)
-        for operator in self.operators:
-            operator.setAlgo(self)
+        # track reference of algorihtm into operator (keep an eye into best solution)
+        for operator in self._operators:
+            if self._parent is not None:
+                operator.setAlgo(self.getParent())
+            else:
+                operator.setAlgo(self)
 
         # also track reference for policy
-        self.policy.setAlgo(self)
+        if self._parent is not None:
+            self._policy.setAlgo(self.getParent())
+        else:
+            self._policy.setAlgo(self)
 
     def addCallback(self, _callback):
[…]
