+++ macop/solutions/BinarySolution.py
@@ -16,7 +16,7 @@ class BinarySolution(Solution):
         size: {int} -- size of binary array values
         score: {float} -- fitness score value
     """
-    def __init__(self, _data, _size):
+    def __init__(self, data, size):
         """
         Initialize binary solution using specific data
 
@@ -25,26 +25,26 @@ class BinarySolution(Solution):
             size: {int} -- size of binary array values
         """
 
-        self.data = _data
-        self.size = _size
+        self._data = data
+        self._size = size
 
-    def random(self, _validator):
+    def random(self, validator):
         """
         Intialize binary array with use of validator to generate valid random solution
 
         Args:
-            _validator: {function} -- specific function which validates or not a solution
+            validator: {function} -- specific function which validates or not a solution
 
         Returns:
             {BinarySolution} -- new generated binary solution
         """
 
-        self.data = np.random.randint(2, size=self.size)
+        self._data = np.random.randint(2, size=self._size)
 
-        while not self.isValid(_validator):
-            self.data = np.random.randint(2, size=self.size)
+        while not self.isValid(validator):
+            self._data = np.random.randint(2, size=self._size)
 
         return self
 
     def __str__(self):
-        return "Binary solution %s" % (self.data)
+        return "Binary solution %s" % (self._data)
