+++ macop/algorithms/multi/MOEAD.py
@@ -15,14 +15,14 @@ from ..Algorithm import Algorithm
 from .MOSubProblem import MOSubProblem
 
 
-def moEvaluator(_solution, _evaluator, _weights):
+def moEvaluator(solution, evaluator, weights):
 
-    scores = [eval(_solution) for eval in _evaluator]
+    scores = [eval(solution) for eval in evaluator]
 
     # associate objectives scores to solution
-    _solution.scores = scores
+    solution._scores = scores
 
-    return sum([scores[i] for i, w in enumerate(_weights)])
+    return sum([scores[i] for i, w in enumerate(weights)])
 
 
 class MOEAD(Algorithm):
@@ -44,96 +44,96 @@ class MOEAD(Algorithm):
         callbacks: {[Callback]} -- list of Callback class implementation to do some instructions every number of evaluations and `load` when initializing algorithm
     """
     def __init__(self,
-                 _mu,
-                 _T,
-                 _initalizer,
-                 _evaluator,
-                 _operators,
-                 _policy,
-                 _validator,
-                 _maximise=True,
-                 _parent=None):
+                 mu,
+                 T,
+                 initalizer,
+                 evaluator,
+                 operators,
+                 policy,
+                 validator,
+                 maximise=True,
+                 parent=None):
 
         # redefinition of constructor to well use `initRun` method
-        self.initializer = _initalizer
-        self.evaluator = _evaluator
-        self.operators = _operators
-        self.policy = _policy
-        self.validator = _validator
-        self.callbacks = []
+        self._initializer = initalizer
+        self._evaluator = evaluator
+        self._operators = operators
+        self._policy = policy
+        self._validator = validator
+        self._callbacks = []
 
         # by default
-        self.numberOfEvaluations = 0
-        self.maxEvaluations = 0
-        self.nObjectives = len(_evaluator)
+        self._numberOfEvaluations = 0
+        self._maxEvaluations = 0
+        self._nObjectives = len(evaluator)
 
         # other parameters
-        self.parent = _parent  # parent algorithm if it's sub algorithm
+        self._parent = parent  # parent algorithm if it's sub algorithm
 
         #self.maxEvaluations = 0 # by default
-        self.maximise = _maximise
+        self._maximise = maximise
 
         # track reference of algo into operator (keep an eye into best solution)
-        for operator in self.operators:
+        for operator in self._operators:
             operator.setAlgo(self)
 
         # by default track reference for policy
-        self.policy.setAlgo(self)
+        self._policy.setAlgo(self)
 
-        if _mu < _T:
+        if mu < T:
             raise ValueError('`mu` cannot be less than `T`')
 
-        self.mu = _mu
-        self.T = _T
+        self._mu = mu
+        self._T = T
 
         # initialize neighbors for each sub problem
         self.setNeighbors()
 
         weights = []
 
-        if self.nObjectives == 2:
+        if self._nObjectives == 2:
 
-            for i in range(self.mu):
-                angle = math.pi / 2 * i / (self.mu - 1)
+            for i in range(self._mu):
+                angle = math.pi / 2 * i / (self._mu - 1)
                 weights.append([math.cos(angle), math.sin(angle)])
 
-        elif self.nObjectives >= 3:
+        elif self._nObjectives >= 3:
 
             # random weights using uniform
-            for i in range(self.mu):
-                w_i = np.random.uniform(0, 1, self.nObjectives)
+            for i in range(self._mu):
+                w_i = np.random.uniform(0, 1, self._nObjectives)
                 weights.append(w_i / sum(w_i))
         else:
             raise ValueError('Unvalid number of objectives')
 
-        self.weights = weights
+        self._weights = weights
 
-        self.subProblems = []
[…]
