+++ docs/build/html/_sources/macop/macop.algorithms.mono.HillClimberFirstImprovment.rst.txt
@@ -1,0 +1,22 @@
+macop.algorithms.mono.HillClimberFirstImprovment
+================================================
+
+.. automodule:: macop.algorithms.mono.HillClimberFirstImprovment
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      HillClimberFirstImprovment
+   
+   
+
+   
+   
+   

[…]
