### 2020-12-11 31fc43d
add protected attributes convention naming for macop
