### 2020-12-11 9f3aca1
Fix issue into basic checkpoint
