+++ docs/source/conf.py
@@ -23,9 +23,9 @@ copyright = '2020, Jérôme BUISINE'
 author = 'Jérôme BUISINE'
 
 # The short X.Y version
-version = '1.0.2'
+version = '1.0.3'
 # The full version, including alpha/beta/rc tags
-release = 'v1.0.2'
+release = 'v1.0.3'
 
 
 # -- General configuration ---------------------------------------------------
