### 2020-12-14 a47fb48
remove evaluation inside policy but inside algorithm
