### 2020-12-14 dbd4ab5
use of inheritance for solutions; check if sol already computed
