+++ macop/solutions/BinarySolution.py
@@ -24,9 +24,7 @@ class BinarySolution(Solution):
             data: {ndarray} --  array of binary values
             size: {int} -- size of binary array values
         """
-
-        self._data = data
-        self._size = size
+        super().__init__(data, size)
 
     def random(self, validator):
         """
