### 2020-12-23 f5b38b6
add unit test for discrete solutions
