+++ macop/solutions/discrete.py
@@ -22,29 +22,50 @@ class BinarySolution(Solution):
         Args:
             data: {ndarray} --  array of binary values
             size: {int} -- size of binary array values
+
+        >>> from macop.solutions.discrete import BinarySolution
+        >>> data = [0, 1, 0, 1, 1]
+        >>> solution = BinarySolution(data, len(data))
+        >>> sum(solution._data) == 3
+        True
+        >>> solution_copy = solution.clone()
+        >>> all(solution_copy._data == solution._data)
+        True
         """
         super().__init__(np.array(data), size)
 
-    def random(self, validator):
+    @staticmethod
+    def random(validator, size):
         """
         Intialize binary array with use of validator to generate valid random solution
 
         Args:
             validator: {function} -- specific function which validates or not a solution
+            size: {int} -- expected solution size to generate
 
         Returns:
             {BinarySolution} -- new generated binary solution
+
+        Example:
+
+        >>> from macop.solutions.discrete import BinarySolution
+        >>> validator = lambda solution: True if sum(solution._data) > 5 else False
+        >>> solution = BinarySolution.random(validator, 10)
+        >>> sum(solution._data) > 5
+        True
         """
 
-        self._data = np.random.randint(2, size=self._size)
+        data = np.random.randint(2, size=size)
+        solution = BinarySolution(data, size)
 
-        while not self.isValid(validator):
-            self._data = np.random.randint(2, size=self._size)
+        while not validator(solution):
+            data = np.random.randint(2, size=size)
+            solution = BinarySolution(data, size)
 
-        return self
+        return solution
 
     def __str__(self):
-        return "Binary solution %s" % (self._data)
+        return f"Binary solution {self._data}"
 
 
 
@@ -64,29 +85,53 @@ class CombinatoryIntegerSolution(Solution):
         Args:
             data: {ndarray} --  array of binary values
             size: {int} -- size of binary array values
+
+        >>> from macop.solutions.discrete import CombinatoryIntegerSolution
+        >>> import numpy as np
+        >>> data = np.arange(5)
+        >>> solution = CombinatoryIntegerSolution(data, 5)
+        >>> sum(solution._data) == 10
+        True
+        >>> solution_copy = solution.clone()
+        >>> all(solution_copy._data == solution._data)
+        True
         """
         super().__init__(data, size)
 
-    def random(self, validator):
+    @staticmethod
+    def random(validator, size):
         """
         Intialize combinatory integer array with use of validator to generate valid random solution
 
         Args:
             validator: {function} -- specific function which validates or not a solution
+            size: {int} -- expected solution size to generate
 
         Returns:
             {CombinatoryIntegerSolution} -- new generated combinatory integer solution
+
+        Example:
+
+        >>> from macop.solutions.discrete import CombinatoryIntegerSolution
+        >>> validator = lambda solution: True if sum(solution._data) > 5 else False
+        >>> solution = CombinatoryIntegerSolution.random(validator, 5)
+        >>> sum(solution._data) > 5
+        True
         """
 
-        self._data = np.random.shuffle(np.arange(self._size))
+        data = np.arange(size)
+        np.random.shuffle(data)
+        solution = CombinatoryIntegerSolution(data, size)
 
-        while not self.isValid(validator):
-            self._data = np.random.shuffle(np.arange(self._size))
+        while not validator(solution):
+            data = np.arange(size)
+            np.random.shuffle(data)
+            solution = CombinatoryIntegerSolution(data, size)
 
-        return self
+        return solution
 
     def __str__(self):
-        return "Combinatory integer solution %s" % (self._data)
+        return f"Combinatory integer solution {self._data}"
 
 
 class IntegerSolution(Solution):
[…]
