+++ docs/build/html/_sources/macop/macop.solutions.continuous.rst.txt
@@ -1,16 +1,0 @@
-macop.solutions.continuous
-==========================
-
-.. automodule:: macop.solutions.continuous
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
