+++ docs/build/html/_sources/macop/macop.evaluators.knapsacks.rst.txt
@@ -1,23 +1,0 @@
-macop.evaluators.knapsacks
-==========================
-
-.. automodule:: macop.evaluators.knapsacks
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      KnapsackEvaluator
-      KnapsackMultiEvaluator
-   
-   
-
-   
-   
-   
