+++ docs/build/html/_sources/macop.rst.txt
@@ -1,60 +1,0 @@
-Documentation
-=============
-
-macop
--------------------
-
-macop.algorithms
--------------------
-
-.. autosummary::
-   :toctree: macop
-   
-   macop.algorithms.base
-   macop.algorithms.mono
-   macop.algorithms.multi
-
-macop.callbacks
--------------------
-
-.. autosummary::
-   :toctree: macop
-   
-   macop.callbacks.base
-   macop.callbacks.classicals
-   macop.callbacks.multi
-   macop.callbacks.policies
-
-macop.evaluators
--------------------
-
-.. autosummary::
-   :toctree: macop
-   
-   macop.evaluators
-   macop.evaluators.knapsacks
-
-macop.operators
--------------------
-
-.. autosummary::
-   :toctree: macop
-   
-   macop.operators.base
-   
-   macop.operators.discrete.mutators
-   macop.operators.discrete.crossovers
-
-   macop.operators.continuous.mutators
-   macop.operators.continuous.crossovers
-
-
-macop.solution
--------------------
-
-.. autosummary::
-   :toctree: macop
-
-   macop.solutions.base
-   macop.solutions.discrete
-   macop.solutions.continuous
