+++ docs/build/html/_sources/macop/macop.callbacks.multi.rst.txt
@@ -1,23 +1,0 @@
-macop.callbacks.multi
-=====================
-
-.. automodule:: macop.callbacks.multi
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      MultiCheckpoint
-      ParetoCheckpoint
-   
-   
-
-   
-   
-   

[…]
