### 2020-12-23 e41526e
add of some doctests
