### 2020-12-24 ed1cb1c
example update for alrogithms and evaluators
