+++ macop/operators/discrete/crossovers.py
@@ -25,7 +25,7 @@ class SimpleCrossover(Crossover):
     >>> from macop.solutions.discrete import BinarySolution
     >>> from macop.algorithms.mono import IteratedLocalSearch
     >>> # evaluator import
-    >>> from macop.evaluators.knapsacks import KnapsackEvaluator
+    >>> from macop.evaluators.discrete.mono import KnapsackEvaluator
     >>> # evaluator initialization (worths objects passed into data)
     >>> worths = [ random.randint(0, 20) for i in range(10) ]
     >>> evaluator = KnapsackEvaluator(data={'worths': worths})
@@ -94,7 +94,7 @@ class RandomSplitCrossover(Crossover):
     >>> from macop.solutions.discrete import BinarySolution
     >>> from macop.algorithms.mono import IteratedLocalSearch
     >>> # evaluator import
-    >>> from macop.evaluators.knapsacks import KnapsackEvaluator
+    >>> from macop.evaluators.discrete.mono import KnapsackEvaluator
     >>> # evaluator initialization (worths objects passed into data)
     >>> worths = [ random.randint(0, 20) for i in range(10) ]
     >>> evaluator = KnapsackEvaluator(data={'worths': worths})
