+++ macop/algorithms/mono.py
@@ -25,6 +25,37 @@ class HillClimberFirstImprovment(Algorithm):
         currentSolution: {Solution} -- current solution managed for current evaluation
         bestSolution: {Solution} -- best solution found so far during running algorithm
         callbacks: {[Callback]} -- list of Callback class implementation to do some instructions every number of evaluations and `load` when initializing algorithm
+    
+    Example:
+
+    >>> import random
+    >>> # operators import
+    >>> from macop.operators.discrete.crossovers import SimpleCrossover
+    >>> from macop.operators.discrete.mutators import SimpleMutation
+    >>> # policy import
+    >>> from macop.policies.classicals import RandomPolicy
+    >>> # solution and algorithm
+    >>> from macop.solutions.discrete import BinarySolution
+    >>> from macop.algorithms.mono import HillClimberFirstImprovment
+    >>> # evaluator import
+    >>> from macop.evaluators.discrete.mono import KnapsackEvaluator
+    >>> # evaluator initialization (worths objects passed into data)
+    >>> problem_size = 20
+    >>> worths = [ random.randint(0, 20) for i in range(problem_size) ]
+    >>> evaluator = KnapsackEvaluator(data={'worths': worths})
+    >>> # validator specification (based on weights of each objects)
+    >>> weights = [ random.randint(5, 30) for i in range(problem_size) ]
+    >>> validator = lambda solution: True if sum([weights[i] for i, value in enumerate(solution._data) if value == 1]) < 200 else False
+    >>> # initializer function with lambda function
+    >>> initializer = lambda x=20: BinarySolution.random(x, validator)
+    >>> # operators list with crossover and mutation
+    >>> operators = [SimpleCrossover(), SimpleMutation()]
+    >>> policy = RandomPolicy(operators)
+    >>> algo = HillClimberFirstImprovment(initializer, evaluator, operators, policy, validator, maximise=True, verbose=False)
+    >>> # run the algorithm
+    >>> solution = algo.run(100)
+    >>> solution._score
+    128
     """
     def run(self, evaluations):
         """
@@ -93,6 +124,37 @@ class HillClimberBestImprovment(Algorithm):
         currentSolution: {Solution} -- current solution managed for current evaluation
         bestSolution: {Solution} -- best solution found so far during running algorithm
         callbacks: {[Callback]} -- list of Callback class implementation to do some instructions every number of evaluations and `load` when initializing algorithm
+    
+    Example:
+
+    >>> import random
+    >>> # operators import
+    >>> from macop.operators.discrete.crossovers import SimpleCrossover
+    >>> from macop.operators.discrete.mutators import SimpleMutation
+    >>> # policy import
+    >>> from macop.policies.classicals import RandomPolicy
+    >>> # solution and algorithm
+    >>> from macop.solutions.discrete import BinarySolution
+    >>> from macop.algorithms.mono import HillClimberBestImprovment
+    >>> # evaluator import
+    >>> from macop.evaluators.discrete.mono import KnapsackEvaluator
+    >>> # evaluator initialization (worths objects passed into data)
+    >>> problem_size = 20
+    >>> worths = [ random.randint(0, 20) for i in range(problem_size) ]
+    >>> evaluator = KnapsackEvaluator(data={'worths': worths})
+    >>> # validator specification (based on weights of each objects)
+    >>> weights = [ random.randint(5, 30) for i in range(problem_size) ]
+    >>> validator = lambda solution: True if sum([weights[i] for i, value in enumerate(solution._data) if value == 1]) < 200 else False
+    >>> # initializer function with lambda function
+    >>> initializer = lambda x=20: BinarySolution.random(x, validator)
+    >>> # operators list with crossover and mutation
[…]
