+++ docs/build/html/_sources/macop/macop.operators.base.rst.txt
@@ -1,25 +1,0 @@
-macop.operators.base
-====================
-
-.. automodule:: macop.operators.base
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      Crossover
-      KindOperator
-      Mutation
-      Operator
-   
-   
-
-   
-   
-   
