+++ docs/build/html/_sources/macop/macop.callbacks.classicals.rst.txt
@@ -1,22 +1,0 @@
-macop.callbacks.classicals
-==========================
-
-.. automodule:: macop.callbacks.classicals
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      BasicCheckpoint
-   
-   
-
-   
-   
-   

[…]
