+++ docs/build/html/_sources/examples.rst.txt
@@ -1,0 +1,626 @@
+Some examples
+=====================================
+
+1. Some information
+--------------------
+
+The package consists of several modules:
+
+- **algorithms**: generic and implemented OR algorithms.
+- **evaluator**: contains all the implemented evaluation functions.
+- **solutions**: all declared solutions classes used to represent problem data.
+- **operators**: mutators, crossovers update of solution. This module also has policies classes to manage the way of update and use solution.
+- **callbacks**: contains Callback class implemenration for making callback instructions every number of evaluations.
+
+
+1.1 Implemented algorithms
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Both single and multi-objective algorithms have been implemented for demonstration purposes. The mono-objective Iterated Local Search algorithm which aims to perform local searches and then to explore again (explorations vs. exploitation trade-off). On the multi-objective side, the MOEA/D algorithm has been implemented using the weighted-sum of objectives method (Tchebycheff approach can also be used). This algorithm aims at decomposing the multi-objective problem into `mu` single-objective problems in order to obtain the pareto front.
+
+1.2 Available solutions
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Currently, only combinatorial solutions are offered, with the well-known knapsack problem as an example. Of course, it's easy to add your own representations of solutions.
+
+1.3 Update solutions
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+A few mutation and crossover operators have been implemented, however it remains quite simple. What is interesting here is that it is possible to develop one's own strategy for choosing operators for the next evaluation. The available UCBPolicy class proposes this functionality as an example, since it will seek to propose the best operator to apply based on a method known as Adaptive Operator Selection (AOS) via the use of the Upper Confidence Bound (UCB) algorithm. 
+
+1.4 Backup feature
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The use of callback instance allows both to save every `k` evaluations of the information but also to reload them once the run of the algorithm is cut. Simply inherit the abstract Callback class and implement the **apply** method to backup and **load** to restore. It is possible to add as many callbacks as required.
+
+
+2. Mono-objective
+-----------------------
+
+In this tutorial, we introduce the way of using `macop` and running your algorithm quickly.
+First of all we need to define the kind of solution which best represent the problem. As example, we use the well known knapsack problem using 30 objects (solution size of 30).
+
+2.1 Problem definition
+~~~~~~~~~~~~~~~~~~~~~~
+
+Hence, we define our problem :
+
+- value of each component of knapsack
+
+- weight associated to each of these components (objects)
+
+.. code:: python
+    
+    """
+    imports part
+    """
+    import random
+
+    """
+    Problem definition
+    """
+    random.seed(42)
+
+    elements_score = [ random.randint(1, 20) for _ in range(30) ] # value of each object
+    elements_weight = [ random.randint(5, 25) for _ in range(30) ] # weight of each object
+
+We can now define the solution representation. In knapsack problem we want to fill our knapsack in an optimisation way selecting or not each component (object).
+The best way to represent this problem is to use the `BinarySolution` from `macop` which stores solution as a binary array.
+
+Using the solution representation, we need to define multiple elements to fit our algorithm :
+
+- function which validates or not a solution (based on constraints)
+
+- function which evaluates the solution (in order to obtain fitness)
+
+- initialization solution function
+
[…]
