+++ docs/build/html/_sources/macop/macop.utils.progress.rst.txt
@@ -1,0 +1,30 @@
+macop.utils.progress
+====================
+
+.. automodule:: macop.utils.progress
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      macop_line
+      macop_progress
+      macop_text
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      Colors
+   
+   
+
+   
+   
+   
