+++ docs/build/html/_sources/macop/macop.solutions.discrete.rst.txt
@@ -1,0 +1,24 @@
+macop.solutions.discrete
+========================
+
+.. automodule:: macop.solutions.discrete
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      BinarySolution
+      CombinatoryIntegerSolution
+      IntegerSolution
+   
+   
+
+   
+   
+   
