+++ docs/build/html/_sources/macop/macop.solutions.base.rst.txt
@@ -1,0 +1,22 @@
+macop.solutions.base
+====================
+
+.. automodule:: macop.solutions.base
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      Solution
+   
+   
+
+   
+   
+   
