+++ docs/build/html/_sources/macop/macop.policies.reinforcement.rst.txt
@@ -1,0 +1,22 @@
+macop.policies.reinforcement
+============================
+
+.. automodule:: macop.policies.reinforcement
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      UCBPolicy
+   
+   
+
+   
+   
+   
