+++ docs/source/api.rst
@@ -1,4 +1,4 @@
-Documentation
+API
 =============
 
 macop
