+++ docs/source/documentations/index.rst
@@ -1,0 +1,18 @@
+Documentation
+==============
+
+.. image:: ../_static/logo_macop.png
+   :width: 300 px
+   :align: center
+
+This documentation will allow a user who wishes to use the Macop optimisation package to understand both how it works and offers examples of how to implement specific needs.
+
+It will gradually take up the major ideas developed within `Macop` to allow for rapid development. You can navigate directly via the menu available below to access a specific part of the documentation.
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Contents:
+
+   introduction
+   problem
+   solutions
