+++ docs/source/documentations/solutions.rst
@@ -1,0 +1,3 @@
+3. Solutions
+=============
+
