+++ docs/build/html/_sources/macop/macop.algorithms.mono.rst.txt
@@ -1,24 +1,0 @@
-macop.algorithms.mono
-=====================
-
-.. automodule:: macop.algorithms.mono
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      HillClimberBestImprovment
-      HillClimberFirstImprovment
-      IteratedLocalSearch
-   
-   
-
-   
-   
-   
