+++ docs/source/documentations/solutions.rst
@@ -1,3 +1,10 @@
 3. Solutions
 =============
 
+A solution is a means of representing a solution to a specific problem in the form of data. In this example, we will always use the knapsack problem as a basis.
+
+3.1. Generic Solution
+~~~~~~~~~~~~~~~~~~~~~~~~~
+
+3.2. Solution representation for knapsack
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
