+++ docs/build/html/_sources/documentations/solutions.rst.txt
@@ -1,3 +1,0 @@
-3. Solutions
-=============
-
