+++ docs/build/html/_sources/macop/macop.evaluators.discrete.multi.rst.txt
@@ -1,22 +1,0 @@
-macop.evaluators.discrete.multi
-===============================
-
-.. automodule:: macop.evaluators.discrete.multi
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      WeightedSum
-   
-   
-
-   
-   
-   
