+++ docs/build/html/_sources/api.rst.txt
@@ -1,79 +1,0 @@
-API
-=============
-
-macop
--------------------
-
-macop.algorithms
--------------------
-
-.. autosummary::
-   :toctree: macop
-   
-   macop.algorithms.base
-   macop.algorithms.mono
-   macop.algorithms.multi
-
-macop.callbacks
--------------------
-
-.. autosummary::
-   :toctree: macop
-   
-   macop.callbacks.base
-   macop.callbacks.classicals
-   macop.callbacks.multi
-   macop.callbacks.policies
-
-macop.evaluators
--------------------
-
-.. autosummary::
-   :toctree: macop
-   
-   macop.evaluators.base
-   macop.evaluators.discrete.mono
-   macop.evaluators.discrete.multi
-
-macop.operators
--------------------
-
-.. autosummary::
-   :toctree: macop
-   
-   macop.operators.base
-   
-   macop.operators.discrete.mutators
-   macop.operators.discrete.crossovers
-
-   macop.operators.continuous.mutators
-   macop.operators.continuous.crossovers
-
-macop.policies
--------------------
-
-.. autosummary::
-   :toctree: macop
-   
-   macop.policies.base
-   
-   macop.policies.classicals
-   macop.policies.reinforcement
-
-macop.solution
--------------------
-
-.. autosummary::
-   :toctree: macop
-
-   macop.solutions.base
-   macop.solutions.discrete
-   macop.solutions.continuous
-
-macop.utils
--------------------
-
-.. autosummary::
-   :toctree: macop
-
-   macop.utils.progress
