+++ docs/build/html/_sources/macop/macop.policies.classicals.rst.txt
@@ -1,22 +1,0 @@
-macop.policies.classicals
-=========================
-
-.. automodule:: macop.policies.classicals
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      RandomPolicy
-   
-   
-
-   
-   
-   
