+++ docs/build/html/_sources/index.rst.txt
@@ -1,28 +1,0 @@
-Minimalist And Customisable Optimisation Package
-================================================
-
-.. image:: _static/logo_macop.png
-   :width: 300 px
-   :align: center
-
-What's `macop` ?
-=================
-
-`macop` is an optimisation Python package which not implement the whole available algorithms in the literature but let you the possibility to quickly develop and test your own algorithm and strategies. The main objective of this package is to be the most flexible as possible and hence, to offer a maximum of implementation possibilities.
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Contents:
-
-   description
-   api
-   documentations/index
-   examples
-   contributing
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
