### 2020-12-29 eda6306
use of classical sphinx command
