### 2021-01-04 063a42c
use of pip for sphinx install
