+++ .github/workflows/docs.yml
@@ -11,9 +11,7 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v2
-      - run: |
-          python -m pip install --upgrade pip
-          pip install -U sphinx
+      - run: sudo apt-get install python3-sphinx
       - run: cd docs && make html
       - name: Deploy
         uses: crazy-max/ghaction-github-pages@v1
