### 2021-01-04 870c7e1
install sphinc theme
