+++ .github/workflows/docs.yml
@@ -11,9 +11,9 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v2
-      - run: python -m pip install --upgrade pip
-      - run: pip install -r requirements.txt
-      - run: pip install sphinx_rtd_theme
+      - run: python3 -m pip install --upgrade pip
+      - run: python3 -m pip install -r requirements.txt
+      - run: python3 -m pip install sphinx_rtd_theme
       - run: cd docs && make html
       - name: Deploy
         uses: crazy-max/ghaction-github-pages@v1
