+++ .github/workflows/docs.yml
@@ -11,6 +11,7 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v2
+      - run: sudo apt install sphinx-build
       - run: python3 -m pip install --upgrade pip
       - run: python3 -m pip install -r requirements.txt
       - run: python3 -m pip install sphinx_rtd_theme
