### 2021-01-04 775a5dd
add sphinx install using apt-get
