### 2021-01-04 0178a39
add sphinx install using apt-get
