+++ .github/workflows/docs.yml
@@ -12,7 +12,6 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v2
       - run: sudo apt-get install python3-sphinx
-      - run: python3 -m pip install --upgrade pip
       - run: python3 -m pip install -r requirements.txt
       - run: python3 -m pip install sphinx_rtd_theme
       - run: cd docs && make html
