+++ .github/workflows/docs.yml
@@ -11,8 +11,9 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v2
-      - run: python3 -m pip install -r requirements.txt
+      - run: python3 -m pip install --user -r requirements.txt
       - run: python3 -m pip install sphinx_rtd_theme
+      - run: export PATH=~/.local/bin/sphinx-build:$PATH
       - run: cd docs && make html
       - name: Deploy
         uses: crazy-max/ghaction-github-pages@v1
