+++ .github/workflows/docs.yml
@@ -13,6 +13,7 @@ jobs:
         uses: actions/checkout@v2
       - run: python3 -m pip install --user -r requirements.txt
       - run: python3 -m pip install sphinx_rtd_theme
+      - run: whereis sphinx-build
       - run: export PATH=~/.local/bin/sphinx-build:$PATH
       - run: cd docs && make html
       - name: Deploy
