+++ .github/workflows/docs.yml
@@ -11,7 +11,7 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v2
-      - run: python3 -m pip install --user -r requirements.txt
+      - run: python3 -m pip install -r requirements.txt
       - run: python3 -m pip install sphinx_rtd_theme
       - run: whereis sphinx-build
       - run: export PATH=/usr/local/bin/sphinx-build:$PATH
