+++ docs/source/conf.py
@@ -18,7 +18,7 @@ sys.path.insert(0, os.path.abspath('../../../macop'))
 
 # -- Project information -----------------------------------------------------
 
-project = 'macop'
+project = 'Macop'
 copyright = '2020, Jérôme BUISINE'
 author = 'Jérôme BUISINE'
 
@@ -90,7 +90,7 @@ language = None
 exclude_patterns = []
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = None
+pygments_style = 'default'
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -98,7 +98,11 @@ pygments_style = None
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
+#html_theme = 'sphinx_rtd_theme'
+#html_theme = 'divio_docs_theme'
+
+html_theme = 'asteroid_sphinx_theme' # use the theme in subdir 'theme'
+# html_theme_path = ['.'] # make sphinx search for themes in current dir
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
