+++ .github/workflows/docs.yml
@@ -15,7 +15,7 @@ jobs:
       with:
         docs-folder: "docs/"
     - name: Deploy
-        uses: crazy-max/ghaction-github-pages@v1
-        with:
-          target_branch: gh-pages
-          build_dir: docs/build
+      uses: crazy-max/ghaction-github-pages@v1
+      with:
+        target_branch: gh-pages
+        build_dir: docs/build
