+++ .github/workflows/docs.yml
@@ -9,13 +9,13 @@ jobs:
   docs:
     runs-on: ubuntu-latest
     steps:
-    - name: Generate
-      uses: actions/checkout@v1
-      uses: ammaraskar/sphinx-action@master
-      with:
-        docs-folder: "docs/"
-    - name: Deploy
-      uses: crazy-max/ghaction-github-pages@v1
-      with:
-        target_branch: gh-pages
-        build_dir: docs/build
+      - name: Generate
+        uses: actions/checkout@v1
+        uses: ammaraskar/sphinx-action@master
+        with:
+          docs-folder: "docs/"
+      - name: Deploy
+        uses: crazy-max/ghaction-github-pages@v1
+        with:
+          target_branch: gh-pages
+          build_dir: docs/build
