### 2021-01-05 e5f011b
solutions documentation added
