+++ .github/workflows/docs.yml
@@ -1,21 +1,38 @@
-name: Package Docs
-on: 
+name: Documentation
+
+on:
   push:
-    branches:
+    branches:    
       - master
       - develop
 
 jobs:
-  docs:
+  build:
     runs-on: ubuntu-latest
     steps:
-      - name: Generate
-        uses: actions/checkout@v1
-        uses: ammaraskar/sphinx-action@master
-        with:
-          docs-folder: "docs/"
-      - name: Deploy
-        uses: crazy-max/ghaction-github-pages@v1
-        with:
-          target_branch: gh-pages
-          build_dir: docs/build
+    - uses: actions/checkout@v1
+    # Standard drop-in approach that should work for most people.
+    - uses: ammaraskar/sphinx-action@master
+      with:
+        docs-folder: "docs/"
+    # Publish built docs to gh-pages branch.
+    # ===============================
+    - name: Commit documentation changes
+      run: |
+        git clone https://github.com/jbuisine/macop.git --branch gh-pages --single-branch gh-pages
+        cp -r docs/build/html/* gh-pages/
+        cd gh-pages
+        touch .nojekyll
+        git config --local user.email "action@github.com"
+        git config --local user.name "GitHub Action"
+        git add .
+        git commit -m "Update documentation" -a || true
+        # The above command will fail if no changes were present, so we ignore
+        # that.
+    - name: Push changes
+      uses: ad-m/github-push-action@master
+      with:
+        branch: gh-pages
+        directory: gh-pages
+        github_token: ${{ secrets.GITHUB_TOKEN }}
+    # ===============================
