+++ .github/workflows/docs.yml
@@ -12,6 +12,7 @@ jobs:
     steps:
     - uses: actions/checkout@v1
     # Standard drop-in approach that should work for most people.
+      run: pip install asteroid-sphinx-theme
     - uses: ammaraskar/sphinx-action@master
       with:
         docs-folder: "docs/"
