+++ .github/workflows/docs.yml
@@ -13,6 +13,7 @@ jobs:
     - uses: actions/checkout@v1
     # Standard drop-in approach that should work for most people.
     - name: Download sphinx theme
+      run: pip3 install asteroid-sphinx-theme
       run: pip install asteroid-sphinx-theme
     - uses: ammaraskar/sphinx-action@master
       with:
