+++ docs/source/conf.py
@@ -14,6 +14,8 @@
 #
 import os
 import sys
+import asteroid_sphinx_theme
+
 sys.path.insert(0, os.path.abspath('../../../macop'))
 
 # -- Project information -----------------------------------------------------
