+++ .github/workflows/docs.yml
@@ -12,11 +12,12 @@ jobs:
     steps:
     - uses: actions/checkout@v1
     # Standard drop-in approach that should work for most people.
-    - name: Download sphinx theme
-      run: pip3 install asteroid-sphinx-theme
     - uses: ammaraskar/sphinx-action@master
       with:
-        docs-folder: "docs/"
+      with:
+        pre-build-command: "pip install asteroid-sphinx-theme"
+        build-command: "make html"
+        docs-folder: "docs"
     # Publish built docs to gh-pages branch.
     # ===============================
     - name: Commit documentation changes
