+++ .github/workflows/docs.yml
@@ -14,7 +14,6 @@ jobs:
     # Standard drop-in approach that should work for most people.
     - uses: ammaraskar/sphinx-action@master
       with:
-      with:
         pre-build-command: "pip install asteroid-sphinx-theme"
         build-command: "make html"
         docs-folder: "docs"
