+++ .github/workflows/docs.yml
@@ -22,7 +22,7 @@ jobs:
     - name: Commit documentation changes
       run: |
         git clone https://github.com/jbuisine/macop.git --branch gh-pages --single-branch gh-pages
-        cp -r docs/build/html/* gh-pages/
+        cp -r docs/* gh-pages/docs
         cd gh-pages
         touch .nojekyll
         git config --local user.email "action@github.com"
