### 2021-01-05 f0203a9
check build content
