### 2021-01-06 52e603a
documentation updates for evaluator / opetators
