+++ macop/algorithms/mono.py
@@ -11,8 +11,8 @@ from .base import Algorithm
 class HillClimberFirstImprovment(Algorithm):
     """Hill Climber First Improvment used as quick exploration optimisation algorithm
 
-    - First, this algorithm do a neighborhood exploration of a new generated solution (by doing operation on the current solution obtained) in order to find a better solution from the neighborhood space.
-    - Then replace the current solution by the first one from the neighbordhood space which is better than the current solution.
+    - First, this algorithm do a neighborhood exploration of a new generated solution (by doing operation on the current solution obtained) in order to find a better solution from the neighborhood space;
+    - Then replace the current solution by the first one from the neighbordhood space which is better than the current solution;
     - And do these steps until a number of evaluation (stopping criterion) is reached.
 
     Attributes:
@@ -110,8 +110,8 @@ class HillClimberFirstImprovment(Algorithm):
 class HillClimberBestImprovment(Algorithm):
     """Hill Climber Best Improvment used as exploitation optimisation algorithm
 
-    - First, this algorithm do a neighborhood exploration of a new generated solution (by doing operation on the current solution obtained) in order to find the best solution from the neighborhood space.
-    - Then replace the best solution found from the neighbordhood space as current solution to use.
+    - First, this algorithm do a neighborhood exploration of a new generated solution (by doing operation on the current solution obtained) in order to find the best solution from the neighborhood space;
+    - Then replace the best solution found from the neighbordhood space as current solution to use;
     - And do these steps until a number of evaluation (stopping criterion) is reached.
 
     Attributes:
@@ -206,12 +206,12 @@ class HillClimberBestImprovment(Algorithm):
 
 
 class IteratedLocalSearch(Algorithm):
-    """Iterated Local Search used to avoid local optima and increave EvE (Exploration vs Exploitation) compromise
+    """Iterated Local Search (ILS) used to avoid local optima and increave EvE (Exploration vs Exploitation) compromise
 
-    - A number of evaluations (`ls_evaluations`) is dedicated to local search process, here `HillClimberFirstImprovment` algorithm
-    - Starting with the new generated solution, the local search algorithm will return a new solution
-    - If the obtained solution is better than the best solution known into `IteratedLocalSearch`, then the solution is replaced
-    - Restart this process until stopping critirion (number of expected evaluations)
+    - A number of evaluations (`ls_evaluations`) is dedicated to local search process, here `HillClimberFirstImprovment` algorithm;
+    - Starting with the new generated solution, the local search algorithm will return a new solution;
+    - If the obtained solution is better than the best solution known into `IteratedLocalSearch`, then the solution is replaced;
+    - Restart this process until stopping critirion (number of expected evaluations).
 
     Attributes:
         initalizer: {function} -- basic function strategy to initialize solution
@@ -222,6 +222,7 @@ class IteratedLocalSearch(Algorithm):
         maximise: {bool} -- specify kind of optimisation problem 
         currentSolution: {Solution} -- current solution managed for current evaluation
         bestSolution: {Solution} -- best solution found so far during running algorithm
+        localSearch: {Algorithm} -- current local search into ILS
         callbacks: {[Callback]} -- list of Callback class implementation to do some instructions every number of evaluations and `load` when initializing algorithm
     
     Example:
@@ -235,6 +236,7 @@ class IteratedLocalSearch(Algorithm):
     >>> # solution and algorithm
     >>> from macop.solutions.discrete import BinarySolution
     >>> from macop.algorithms.mono import IteratedLocalSearch
+    >>> from macop.algorithms.mono import HillClimberFirstImprovment
     >>> # evaluator import
     >>> from macop.evaluators.discrete.mono import KnapsackEvaluator
[…]
