+++ docs/source/conf.py
@@ -25,9 +25,9 @@ copyright = '2020, Jérôme BUISINE'
 author = 'Jérôme BUISINE'
 
 # The short X.Y version
-version = '1.0.5'
+version = '1.0.6'
 # The full version, including alpha/beta/rc tags
-release = 'v1.0.5'
+release = 'v1.0.6'
 
 
 # -- General configuration ---------------------------------------------------
