+++ docs/source/documentations/introduction.rst
@@ -1,7 +1,7 @@
 1. Introduction
 ================
 
-`Macop` is a python package for solving optimisation problems that are both discrete and continuous in nature. The objective is to allow a user to exploit the basic structure proposed by this package to solve a problem specific to him. The interest is that he can quickly abstract himself from the complications related to the way of evaluating, comparing, saving the progress of the search for good solutions but rather concentrate if necessary on his own algorithm. Indeed, `Macop` offers the following main and basic features: 
+`Macop` is a python package for solving discrete optimisation problems in nature. Continuous optimisation is also applicable but not yet developed. The objective is to allow a user to exploit the basic structure proposed by this package to solve a problem specific to him. The interest is that he can quickly abstract himself from the complications related to the way of evaluating, comparing, saving the progress of the search for good solutions but rather concentrate if necessary on his own algorithm. Indeed, `Macop` offers the following main and basic features: 
 
 - **solutions:** representation of the solution ;
 - **validator:** such as constraint programmig, a `validator` is function which is used for validate or not a solution data state ;
