### 2021-01-06 d2fb195
End of documentation for operators selection process
