+++ macop/policies/base.py
@@ -4,6 +4,9 @@ import logging
 from abc import abstractmethod
 
 
+from ..operators.base import KindOperator
+
+
 # define policy to choose `operator` function at current iteration
 class Policy():
     """Abstract class which is used for applying strategy when selecting and applying operator 
@@ -24,12 +27,13 @@ class Policy():
         """
         pass
 
-    def apply(self, solution):
+    def apply(self, solution1, solution2=None):
         """
-        Apply specific operator chosen to create new solution, computes its fitness and returns solution
+        Apply specific operator chosen to create new solution, compute its fitness and return solution
         
         Args:
-            _solution: {Solution} -- the solution to use for generating new solution
+            solution1: {Solution} -- the first solution to use for generating new solution
+            solution2: {Solution} -- the second solution to use for generating new solution (in case of specific crossover, default is best solution from algorithm)
 
         Returns:
             {Solution} -- new generated solution
@@ -38,12 +42,25 @@ class Policy():
         operator = self.select()
 
         logging.info("---- Applying %s on %s" %
-                     (type(operator).__name__, solution))
+                     (type(operator).__name__, solution1))
+
+        # default value of solution2 is current best solution
+        if solution2 is None and self._algo is not None:
+            solution2 = self._algo._bestSolution
+
+        # avoid use of crossover if only one solution is passed
+        if solution2 is None and operator._kind == KindOperator.CROSSOVER:
+
+            while operator._kind == KindOperator.CROSSOVER:            
+                operator = self.select()
 
         # apply operator on solution
-        newSolution = operator.apply(solution)
+        if operator._kind == KindOperator.CROSSOVER:
+            newSolution = operator.apply(solution1, solution2)
+        else:
+            newSolution = operator.apply(solution1)
 
-        logging.info("---- Obtaining %s" % (solution))
+        logging.info("---- Obtaining %s" % (newSolution))
 
         return newSolution
 
