### 2021-01-07 6c01237
add of algorithms documentation
