+++ docs/source/documentations/algorithms.rst
@@ -1,2 +1,364 @@
-8. Optimisation process
+Optimisation process
 =======================
+
+Let us now tackle the interesting part concerning the search for optimum solutions in our research space.
+
+Find local and global optima
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Overall, in an optimization process, we will seek to find the best, or the best solutions that minimize or maximize our objective function (fitness score obtained) in order to respond to our problem.
+
+.. image:: ../_static/documentation/search_space.png
+   :width:  800 px
+   :align: center
+
+Sometimes, the search space can be very simple. A local search can provide access to the global optimum as shown in figure (a) above. 
+In other cases, the search space is more complex. It may be necessary to explore more rather than exploit in order to get out of a convex zone and not find the global optimum but only a local opmatime solution. 
+This problem is illustrated in figure (b).
+
+Abstract algorithm class
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+An abstract class is proposed within Macop to generalize the management of an algorithm and therefore of a heuristic. 
+It is located in the ``macop.algorithms.base`` module. 
+
+We will pay attention to the different methods of which she is composed. This class enables to manage some common usages of operation research algorithms:
+
+- initialization function of solution
+- validator function to check if solution is valid or not (based on some criteria)
+- evaluation function to give fitness score to a solution
+- operators used in order to update solution during search process
+- policy process applied when choosing next operator to apply
+- callbacks function in order to do some relative stuff every number of evaluation or reload algorithm state
+- parent algorithm associated to this new algorithm instance (hierarchy management)
+
+She is composed of few default attributes:
+
+- initializer: {function} -- basic function strategy to initialize solution
+- evaluator: {Evaluator} -- evaluator instance in order to obtained fitness (mono or multiple objectives)
+- operators: {[Operator]} -- list of operator to use when launching algorithm
+- policy: {Policy} -- Policy instance strategy to select operators
+- validator: {function} -- basic function to check if solution is valid or not under some constraints
+- maximise: {bool} -- specify kind of optimisation problem 
+- verbose: {bool} -- verbose or not information about the algorithm
+- currentSolution: {Solution} -- current solution managed for current evaluation comparison
+- bestSolution: {Solution} -- best solution found so far during running algorithm
+- callbacks: {[Callback]} -- list of Callback class implementation to do some instructions every number of evaluations and `load` when initializing algorithm
+- parent: {Algorithm} -- parent algorithm reference in case of inner Algorithm instance (optional)
+
+.. code-block:: python
+
+    class Algorithm():
+
+        def __init__(self,
+                    initializer,
+                    evaluator,
+                    operators,
+                    policy,
+                    validator,
+                    maximise=True,
+                    parent=None,
+                    verbose=True):
+            ...
+
+        def addCallback(self, callback):
+            """
+            Add new callback to algorithm specifying usefull parameters
+            """
+            ...
+
+        def resume(self):
+            """
+            Resume algorithm using Callback instances
+            """
+            ...
+
+        def getParent(self):
+            """
+            Recursively find the main parent algorithm attached of the current algorithm
+            """
+            ...
+
+        def setParent(self, parent):
+            """
+            Set parent algorithm to current algorithm
+            """
+            ...
+
+
+        def initRun(self):
+            """
+            Initialize the current solution and best solution using the `initialiser` function
+            """
+            ...
+
+        def increaseEvaluation(self):
+            """
+            Increase number of evaluation once a solution is evaluated for each dependant algorithm (parents hierarchy)
+            """
+            ...
+                
+        def getGlobalEvaluation(self):
[…]

[…]
