+++ docs/source/documentations/callbacks.rst
@@ -1,2 +1,233 @@
 Keep track
 ==============
+
+Keeping track of the running algorithm can be useful on two levels. First of all to understand how it unfolded at the end of the classic run. But also in the case of the unwanted shutdown of the algorithm. 
+This section will allow you to introduce the recovery of the algorithm thanks to a continuous backup functionality.
+
+Logging into algorithm
+~~~~~~~~~~~~~~~~~~~~~~
+
+Some logs can be retrieve after running an algorithm. **Macop** uses the ``logging`` Python package in order to log algorithm advancement.
+
+Here is an example of use when running an algorithm:
+
+.. code-block:: python
+
+    """
+    basic imports
+    """
+    import logging
+
+    # logging configuration
+    logging.basicConfig(format='%(asctime)s %(message)s', filename='data/example.log', level=logging.DEBUG)
+
+    ...
+    
+    # maximizing algorithm (relative to knapsack problem)
+    algo = HillClimberBestImprovment(initialiser, evaluator, operators, policy, validator, maximise=True, verbose=False)
+
+    # run the algorithm using local search and get solution found 
+    solution = algo.run(evaluations=100)
+    print(solution.fitness())
+
+Hence, log data are saved into ``data/example.log`` in our example.
+
+Callbacks introduction
+~~~~~~~~~~~~~~~~~~~~~~~
+
+Having output logs can help to understand an error that has occurred, however all the progress of the research carried out may be lost. 
+For this, the functionality relating to callbacks has been developed.
+
+Within **Macop**, a callback is a specific instance of ``macop.callbacks.Callback`` that allows you to perform an action of tracing / saving information **every** ``n`` **evaluations** but also reloading information if necessary when restarting an algorithm.
+
+
+.. code-block:: python
+
+    class Callback():
+
+        def __init__(self, every, filepath):
+            ...
+
+        @abstractmethod
+        def run(self):
+            """
+            Check if necessary to do backup based on `every` variable
+            """
+            pass
+
+        @abstractmethod
+        def load(self):
+            """
+            Load last backup line of solution and set algorithm state at this backup
+            """
+            pass
+
+        def setAlgo(self, algo):
+            """
+            Specify the main algorithm instance reference
+            """
+            ...
+
+
+- The ``run`` method will be called during run process of the algo and do backup at each specific number of evaluations. 
+- The ``load`` method will be used to reload the state of the algorithm from the last information saved. All saved data is saved in a file whose name will be specified by the user.
+
+Towards the use of Callbacks
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+We are going to create our own Callback instance called ``BasicCheckpoint`` which will save the best solution found and number of evaluations done in order to reload it for the next run of our algorithm.
+
+.. code-block:: python
+
+    """
+    module imports
+    """
+    from macop.callbacks.base import Callback
+
+
+    class BasicCheckpoint(Callback):
+        
+        def run(self):
+            """
+            Check if necessary to do backup based on `every` variable
+            """
+            # get current best solution
+            solution = self._algo._bestSolution
+
+            currentEvaluation = self._algo.getGlobalEvaluation()
+
+            # backup if necessary every number of evaluations
+            if currentEvaluation % self._every == 0:
+
+                # create specific line with solution data
+                solutionData = ""
+                solutionSize = len(solution._data)
+
+                for index, val in enumerate(solution._data):
+                    solutionData += str(val)
+
+                    if index < solutionSize - 1:
+                        solutionData += ' '
+
+                # number of evaluations done, solution data and fitness score
+                line = str(currentEvaluation) + ';' + solutionData + ';' + str(
+                    solution.fitness()) + ';\n'
+
+                # check if file exists
+                if not os.path.exists(self._filepath):
+                    with open(self._filepath, 'w') as f:
[…]
