+++ paper.bib
@@ -132,3 +132,51 @@
   biburl    = {https://dblp.org/rec/journals/remotesensing/PullanagariKY18.bib},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
+
+@misc{ceres-solver,
+  author = "Sameer Agarwal and Keir Mierle and Others",
+  title = "Ceres Solver",
+  howpublished = "\url{http://ceres-solver.org}",
+}
+
+@book{hart2017pyomo,
+  title={Pyomo--optimization modeling in python},
+  author={Hart, William E. and Laird, Carl D. and Watson, Jean-Paul and Woodruff, David L. and Hackebeil, Gabriel A. and Nicholson, Bethany L. and Siirola, John D.},
+  edition={Second},
+  volume={67},
+  year={2017},
+  publisher={Springer Science \& Business Media}
+}
+
+@article{pyopt-paper,
+  author = {Ruben E. Perez and Peter W. Jansen and Joaquim R. R. A. Martins},
+  title = {py{O}pt: A {P}ython-Based Object-Oriented Framework for Nonlinear Constrained Optimization},
+  journal = {Structures and Multidisciplinary Optimization},
+  year = {2012},
+  volume = {45},
+  number = {1},
+  pages = {101--118},
+  doi = {10.1007/s00158-011-0666-3}
+}
+
+@incollection{MaherMiltenbergerPedrosoRehfeldtSchwarzSerrano2016,
+  author = {Stephen Maher and Matthias Miltenberger and Jo{\~{a}}o Pedro Pedroso and Daniel Rehfeldt and Robert Schwarz and Felipe Serrano},
+  title = {{PySCIPOpt}: Mathematical Programming in Python with the {SCIP} Optimization Suite},
+  booktitle = {Mathematical Software {\textendash} {ICMS} 2016},
+  publisher = {Springer International Publishing},
+  pages = {301--307},
+  year = {2016},
+  doi = {10.1007/978-3-319-42432-3_37},
+}
+
+@misc{simanneal-solver,
+  author = "Matthew Perry",
+  title = "simanneal",
+  howpublished = "\url{https://github.com/perrygeo/simanneal}",
+}
+
+@misc{solid-solver,
+  author = "Devin Soni",
+  title = "Solid",
+  howpublished = "\url{https://github.com/100/Solid}",
+}
