### 2021-01-07 2b1151e
update of references into paper and figure
