### 2021-01-07 #2812 @jbuisine
Hello @melissawm, @stsievert, @torressa

First of all, I would like to apologise for the delay in responding to your comments. I would like to thank you for your very good remarks, I have modified the document (paper), the package and the documentation accordingly. Please find below my comments in relation to your various questions/comments.

The comments to the different questions are answered for each reviewer in two sections:

1. the answers arising from the reviewer's comments
2. the answers to the remarks in the JOSS Journal Specification Grid (or requirement box not checked).

According to a certain recommendation, the answer to reviewers is presented as follows:

- **[RV]** Reviewer Comments
- **[AA]** Authors' Answers

---

# @stsievert

## 1. Most significant comments/questions

### 1.1. Tests
**[RV]** The tests are not sufficient. I only found a single test in setup.py, and not a single doctest as claimed. In this one test, there are many untested classes.

**[AA]** Updates have been done following this comment using doctest in order to better cover the code.

### 1.2. Audience
**[RV]** Target audience and statement of need. The main documentation do not identify a target audience or a statement of need. Only one feature of the software, maximum flexibility, is described. Who would want that and why? In paper.md, I see that there's a need for maximum flexibility mentioned in the paper, but why would a user want if they're not doing "thesis work"?

**[AA]** The proposed Macop package is now specified as being oriented towards discrete optimization. It therefore targets anyone wishing to respond to a discrete optimization problem initially.

Macop wants to allow users to quickly focus on one of proposed features from its interaction main loop.

### 1.3. API
**[RV]** The documentation has some specific problems: it's very general and not detailed, is missing a good API description.

- **Generality:** optimization is very general. When is Macop not appropriate?
        For example, could I use it for my favorite deep learning problem? Could it be used for a continuous convex optimization?
        Or what if the optimization problem I'm using requires sparse arrays or GPUs?
- **Detailed:** the only information core classes have is their name – past that, there's not much detail.
        For example, the documentation for IteratedLocalSearch is basically "[this class is] used to avoid local optima and increave EvE (Exploration vs Exploitation) compromise". What's "EvE", what effect does it have, and how do the various parameters balance it?
        For example, what do C and exp_rate do in UCBPolicy? Does increasing C increase or decrease the importance of exploration? How does changing the value of exp_rate change the exploration/exploitation tradeoff?
- **API:** it's not clear what different policies, mutators and crossovers are used for. What's the data flow of one example? Detailing the input/output of these modules would be useful. I'd expect to see a clear definitions for non-experts.

**[AA]** The proposed Macop package is now specified as being oriented towards discrete optimization. However, it remains open to extension for continued optimization. Its main objective is to offer a common interaction loop and to allow users of the package to focus their developments on one of the parts in question. It therefore targets anyone wishing to respond to a discrete optimization problem initially.

All documentation has been updated. It is possible to understand the full features interaction and main behaviour of the package with a complete example. A more complete API is also available.

## 2. edited #2812 (comment)

### Functionality

**[RV]** _Installation:_ Specifically, this claim is made in paper.md: "Solutions modeling continuous problems can also be created by the anyone who wants to model his own problem", but I'm not seeing a continuous optimization problem in the examples or tests.

**[AA]** The proposed Macop package is now specified as being oriented towards discrete optimization. However, it remains open to extension for continued optimization but it was not dealt with.

### Documentation

**[RV]** _A statement of need:_ Do the authors clearly state what problems the software is designed to solve and who the target audience is?

**[AA]** Macop now targets anyone wishing to respond to a discrete optimization problem.

**[RV]** __Functionality documentation__: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

**[AA]** All documentation has been updated. It is possible to understand the full features interaction and main behaviour of the package. A more complete API is also available.

**[RV]** __Automated tests__: Are there automated tests or manual steps described so that the functionality of the software can be verified?

    Yes, but these tests are not sufficient; some modules are untouched. See below for details.

**[AA]** Updates have been done following this comment using doctest in order to better cover the whole code.

**[RV]** __Community guidelines__: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

    Yes, there's some guidelines for contributing code/PRs but it's not super clear.
    Yes, there is some guidelines for filing issues/bug reports but it only says "submit an issue with label ...".
    There really aren't any guidelines for seeking support past "do not hestitate" when contacting.
    I think this could be a lot clearer; I'm leaving this requirement unchecked.

**[AA]** Some improvements have been added in order to make it clearer how it is possible to contribute to the package.

### Software paper

**[RV]** __Summary__: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
        Not really. It claims to solve every optimization problem; that seems a little to general. I don't know what job/data would motivate using this software.

**[AA]** The proposed Macop package is now specified as being oriented towards discrete optimization.

**[RV]** __A statement of need__: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

**[AA]** Details have been made in the paper.md document and also in the package documentation.

**[RV]** __State of the field__: Do the authors describe how this software compares to other commonly-used packages?
        Marginally – I think this requirement is partially satisfied. The entirety of the comparison is "Available libraries in the literature did not allow this kind of implementation of an evaluation function quickly"

**[AA]** Citations have been updated to better situated where Macop.

**[RV]** __Quality of writing__: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
        See below for specific edits.

**[AA]** The document has been updated with respect to the various remarks proposed.

---

# @torressa

## 1. Most significant comments/questions

### 1.1. Documentation

**[RV]**
- The api is present but badly documented (e.g. "Iterated Local Search Algorithm implementation") and not up to date with the master branch. It needs more elaborate explanations of the implementations and citations where appropriate.
- The README and paper do not present an accurate description of what Macop actually is. For this you have to go through the code. Clearer technical description (in paper too)
- The uses of the toolbox has to be better defined, where is it better to use Macop over exact or other methods.

**[AA]** All documentation has been updated. It is possible to understand the full features interaction and main behaviour of the package. A more complete API is also available.

### 1.2. Testing

**[RV]** bad. Unit tests missing (there is one test in the setup.py file)

**[AA]** Updates have been done following this comment using doctest in order to better cover the whole code.

### 1.3. LICENSE

**[RV]** not completed.

**[AA]** Updates have been done in order to well complete the License.

### 1.4. Scholarly effort

Possibly not a substantial scholarly effort

**[RV]**
- I've found the bulk of the algorithmic implementation to be mostly contained in July 2019 and August-October 2020. The rest of the work is re-structuring and packaging. Therefore, raising the question of whether the time spent on this project was greater than 3 months.
- To add to this, as reported by whedon there's 1114 lines of python code. Given the structure of the code (with the nested inheritance and redefinition of functions), it is hard to say whether this ticks the "Substantial scholarly effort" box.

**[AA]** It is true that the development of the package was done over periods rather than on continuity. Updates to the package have also been offered in the current latest version.

### 1.5. Software paper

**[RV]**

- **State of the field:** there is no mention of other similar software packages. The author comments on well-known general optimisation packages (which need a proper citation), but nothing related to the methods used in Macop. In a quick google-search I found a more closely related repo: Solid which provides multiple derivate-free algorithms with user-defined objectives. The author probably knows more about this but I suspect there must be others, either way, it should be stated in the paper.

- **Quality of writing:** the paper needs rereading to correct minor grammatical / orthographical errors. I will provide a list once the other comments have been addressed.

**[AA]** Updates have been done regarding these comments in order to better situate Macop in litterature.

### 1.6. Major Issues (code):

**[RV]**

- **Software usage:** the usage of library is not very clear. How does one extract the final solution without looking at the logs? Not the user defined criteria like the score/fitness, but the final solution (i.e. in the knapsack example, the items in the knapsack).

- Why is abstract class inheritance only sometimes used? This seems like the right type of inheritance for all the classes.

- `macop/evaluators`: contains a single example and not an implementation

- `macop/operators`: triple nested inheritance makes it hard to determine where functions are actually being implemented (e.g. UCBPolicy inherits from Policy which inherits from Operator). For such simple functionality, e.g definining the method apply, this seems a bit overkill. Additionally, inheritance is not used e.g. macop/operators/policies/UCBPolicy.py could be a standalone class.

- `macop/solutions/`: Same as above, inheritance doesn't do much as functions are overloaded and arguments not passed to the parent. The purpose of these functions is really not clear to me.

- In general, the module like structure is really confusing, specially given the small and simple implementations. This makes the imports list to extend a fair bit. I understand that this structure may have been laid out for additional functionality to be introduced. However, if no more work is to be done, to help people understand the code and encourage contributions I suggest a major rework of the code to simplify the structure.

**[AA]** Many updates have been done in term of code:

- Usage of Macop is now better specified in the documentation of software with a full example.
- Inheritance is now always used.
- `macop.evaluators`: is now composed of a class and not just a function. Evaluator is composed of the required data of the problem instance.
- `macop.operators`: Operator is basic abstract class, Mutator and Crossover inherit from Operator. Then implementation of Mutator and Crossover are available. UCBPolicy only inherite from Policy abstract class.
- `macop.solutions`: abstract method is now specified to better understand expected inheritance behaviour.
- A major rework have been done to better structure the whole Package using module instead of single class per file.

### 1.7. Minor comments

**[RV]** Please get rid of the build files for the documentation. readthedocs can build the docs on commits to master if you set it up.

**[AA]** I now use the classical `gh-pages` branch to automatically update the doc using CI.

**[RV]** Relative imports are discouraged by PEP 328 "With the shift to absolute imports, the question arose whether relative imports should be allowed at all".

**[AA]** Updates have been done.

**[RV]** Remove leading _ for all input parameters, this is common practice for private parameters.

**[AA]** Updates have been done. Instance attributes now possessed the leading `_`.

**[RV]** `macop/callbacks`: they are technically a callback, but it seems more like a custom logger

**[AA]** Yes it's true, but it enables to reload the algorithm state if interuption.

**[RV]** Is there any way to disable the progress output being printed to screen? I can imagine this needs disabling in normal applications

**[AA]** `verbose` attribute to algorithm is now enable and disable the progress.

**[RV]**  The knapsack is a nice an simple example, however, I don't think it showcases the best use of this library. As it takes longer to arrive at the optimal solution with 30 elements with Macop than traditional exact methods.

**[AA]** It is true that the knapsack is a very classic example and whose solutions can be quickly found. However, I wanted to keep it as an example as well in the documentation due to its simplicity. I will add examples of other instances of problems later.

**[RV]** In a standard python package, the init file in the main directory is not needed. init file in macop/ should contain module imports

**[AA]** Updates have been done into `__init__.py` main file of `macop`.

**[RV]** Contributions file is really long while explanations on how to contribute are short.

**[AA]** I tried to update it to make it a little more clearer.

**[RV]** Examples docs page is really long (half of it is imports and multi-line comments)

**[AA]** I update the whole documentation with sections to describe the package usages.

**[RV]**In README:
   - Tags in readme should link to ci results and external urls, not images.
   - "OR algorithms"? list them.

**[AA]** Some updates have been done in README.md but not sure it was the expected things.

## 2. edited #2812 (comment)

### General checks

I do not fill this part because all comments are answered above.
