### 2021-01-11 0744273
use of specific version of sphinx for API documentation
