+++ .github/workflows/docs.yml
@@ -14,7 +14,7 @@ jobs:
     # Standard drop-in approach that should work for most people.
     - uses: ammaraskar/sphinx-action@master
       with:
-        pre-build-command: "pip install asteroid-sphinx-theme"
+        pre-build-command: "pip uninstall sphinx && pip install sphinx==v2.2.0 && pip install asteroid-sphinx-theme"
         build-command: "make html"
         docs-folder: "docs"
     # Publish built docs to gh-pages branch.
