### 2021-01-11 4bc1167
enable remove of previous sphinx version
