### 2021-01-19 #2812 @torressa
Hi @jbuisine!
Thanks for the detailed response and the efforts on the package.
I am much happier now with the work you've done so I'm ticking off some of the requirements.
Also, really good work on the documentation and diagrams, it looks really nice.

I'll take some time over the week end to come back with my feedback.
