### 2021-01-24 #2812 @torressa
Hi again,
Here's my resp

### Code

Thanks taking the comments on board and standardising the inheritance structure. The imports look nicer now.
I like that you've included placeholders for future implementation for the continuous case.

Only one more thing:

Is it possible to have a more complex example where you have used macop? I like the knapsack example in the docs, but it would be good to have an advanced usage example. If you are using for something in your research, maybe a minimal working example would be good. I've noticed the noise-detection repo, maybe link that it the Macop README.

### Paper

Minor comments

- Page 1 line 13 repeated "which doesn't"
- Page 1 line 16 repeated full stop ".."
- Page 2 line 22. The list of optimisation packages is not extensive, I suggest to remove it and just cite software which deals with algorithms closer to Macop. The Perry + Soni Citations should be changed to cite a specific version and year (I've checked and both have releases with year).
- Page 2 paragraph starting in page 27 doesn't seem to belong there. The motivation starting in line 35 seems more appropriate to wrap the software discussion.
- Page 2 bullets starting in line 47, remove spaces before the semi-colons.
- Page 2 line 48 "programming"
- Page 2 section beginning on line 61, please refer back to the modules. E.g say in `macop.algorihtms`. Also, I would include this information in the README.
- Page 3 sections "Available solutions", "Update solutions", and "Callback feature", same as above, refer to the module (`macop.solutions`, `macop.evaluators`,...) for completeness.
- Page 4 line 105, I get what you mean by "dynamic programming of the Python..." but it's a bit confusing as dynamic programming also means something completely unrelated. Please rephrase this.
