### 2021-01-24 0ae127a
update of macop paper and README.md
