+++ paper.bib
@@ -172,11 +172,15 @@
 @misc{simanneal-solver,
   author = "Matthew Perry",
   title = "simanneal",
+  year= "2019",
+  version= "0.5.0",
   howpublished = "\url{https://github.com/perrygeo/simanneal}",
 }
 
 @misc{solid-solver,
   author = "Devin Soni",
   title = "Solid",
+  version = "0.11",
+  year = "2017",
   howpublished = "\url{https://github.com/100/Solid}",
 }
