+++ paper.bib
@@ -184,3 +184,21 @@
   year = "2017",
   howpublished = "\url{https://github.com/100/Solid}",
 }
+
+@inproceedings{10.1145/3321707.3321800,
+author = {Lepr\^{e}tre, Florian and Verel, S\'{e}bastien and Fonlupt, Cyril and Marion, Virginie},
+title = {Walsh Functions as Surrogate Model for Pseudo-Boolean Optimization Problems},
+year = {2019},
+isbn = {9781450361118},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3321707.3321800},
+doi = {10.1145/3321707.3321800},
+abstract = {Surrogate-modeling is about formulating quick-to-evaluate mathematical models, to approximate black-box and time-consuming computations or simulation tasks. Although such models are well-established to solve continuous optimization problems, very few investigations regard the optimization of combinatorial structures. These structures deal for instance with binary variables, allowing each compound in the representation of a solution to be activated or not. Still, this field of research is experiencing a sudden renewed interest, bringing to the community fresh algorithmic ideas for growing these particular surrogate models. This article proposes the first surrogate-assisted optimization algorithm (WSaO) based on the mathematical foundations of discrete Walsh functions, combined with the powerful grey-box optimization techniques in order to solve pseudo-boolean optimization problems. We conduct our experiments on a benchmark of combinatorial structures and demonstrate the accuracy, and the optimization efficiency of the proposed model. We finally highlight how Walsh surrogates may outperform the state-of-the-art surrogate models for pseudo-boolean functions.},
+booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference},
+pages = {303–311},
+numpages = {9},
+keywords = {combinatorial optimization, surrogate model/fitness approximation, local search, empirical study},
+location = {Prague, Czech Republic},
+series = {GECCO '19}
+}
