### 2021-01-24 abe6a93
Update of paper and macop version
