+++ README.md
@@ -11,12 +11,12 @@
 
 `Macop` is a python package for solving discrete optimisation problems in nature. Continuous optimisation can also applicable if needed. The objective is to allow a user to exploit the basic structure proposed by this package to solve a problem specific to him. The interest is that he can quickly abstract himself from the complications related to the way of evaluating, comparing, saving the progress of the search for good solutions but rather concentrate if necessary on his own algorithm. Indeed, `Macop` offers the following main and basic features: 
 
-- **solutions:** representation of the solution ;
-- **validator:** such as constraint programmig, a `validator` is function which is used for validate or not a solution data state ;
-- **evaluator:** stores problem instance data and implement a `compute` method in order to evaluate a solution ;
-- **operators:** mutators, crossovers update of solution ;
-- **policies:** the way you choose the available operators (might be using reinforcement learning) ;
-- **algorithms:** generic and implemented optimisation research algorithms ;
+- **solutions:** representation of the solution;
+- **validator:** such as constraint programmig, a `validator` is function which is used for validate or not a solution data state;
+- **evaluator:** stores problem instance data and implement a `compute` method in order to evaluate a solution;
+- **operators:** mutators, crossovers update of solution;
+- **policies:** the way you choose the available operators (might be using reinforcement learning);
+- **algorithms:** generic and implemented optimisation research algorithms;
 - **callbacks:** callbacks to automatically keep track of the search space advancement.
 
 <p align="center">
@@ -68,7 +68,7 @@ git submodule add https://github.com/jbuisine/macop.git
 
 ## Current projects which use `Macop`
 
-- [@prise3d/noise-detection-attributes-optimization](https://github.com/prise-3d/noise-detection-attributes-optimization): use of a parent algorithm with the real (but very expensive) evaluation function, and then local searches use a substitution model (a model that has learned to approximate the real cost function). Hence two evaluation functions have been used in order to accelerate the search in the set of solutions.
+- [@prise3d/noise-detection-attributes-optimization](https://github.com/prise-3d/noise-detection-attributes-optimization): use of a parent algorithm with the real (but very expensive) evaluation function, and then inner local searches which use a substitution model (a model that has learned to approximate the real cost function with a quick-to-evaluate function). Hence, two evaluation functions have been used in order to accelerate the search in the set of solutions.
 
 ## License
 
