### 2021-01-24 #2812 @jbuisine
Hi @toressa,

Thank you very mush for your answer.

## Code

Regarding the code, I really appreciate your comments. I can indeed propose a more complex example (or even several).

I can use QAP (quadratic assignment problem) or UBQP (unconstrained binary quadratic programming) instances. Instances are proposed on [mocobench](http://mocobench.sourceforge.net/), but I may stay for the examples in a single-objective context.

In my work on noise detection, this would remain more complicated to explain, in reality, I use a parent algorithm with the real (but very expensive) evaluation function, and then local searches which use a surrogate model (a model that has learned to approximate the real cost function and with a quick-to-evaluate function). I'm afraid this is too far from the generic aspect of the package. But I specified in the paper the interest of the proposed package for this feature (which I specify later in this comment).

So if it suits you, I can add in the documentation some examples on a QAP and UBQP instance.

## Paper

- Page 1 line 13: fixex
- Page 1 line 16: fixed
- Page 2 line 22: list updated
- Page 2 paragraph starting in line 27: fixed
- Page 2 bullets starting in line 47: fixed
- Page 2 line 48: fixed
- Page 2 section beginning on line 61: fixed
- Page 3 sections "Available solutions", "Update solutions", and "Callback feature": fixed
- Page 4 line 105: changed to "flexible"
- I also added the description in the paper bringing the fact that despite the fact that a dependency exists between different algorithms, each one can have its own context (update operators, evaluation function, solution update policy...). This means that only the results (solutions found) are exchanged for example during the search process.
- README.md have been also updated consequently.

Thank you again for your return and usefull comment.
