### 2021-01-25 #2812 @torressa
Thanks for the fast corrections.

> So if it suits you, I can add in the documentation some examples on a QAP and UBQP instance.

That sounds good, would be a great addition!
