### 2021-01-25 #2812 @melissawm
Hello @stsievert - would you be able to take a second look at the paper now that @jbuisine has responded to comments? Let me know if I can help in any way. Thanks!
