### 2021-01-26 74491c1
add of QAP problem example into documentation
