+++ examples/ubqpExample.py
@@ -23,7 +23,7 @@ if not os.path.exists('data'):
     os.makedirs('data')
 
 # logging configuration
-logging.basicConfig(format='%(asctime)s %(message)s', filename='data/example_knapsack.log', level=logging.DEBUG)
+logging.basicConfig(format='%(asctime)s %(message)s', filename='data/example.log', level=logging.DEBUG)
 
 random.seed(42)
 
