+++ docs/source/contributing.rst
@@ -8,10 +8,6 @@ Contributing
 Using GitHub
 ------------
 
-This git project uses git-flow_ implementation. You are free to contribute to it.
-
-.. _git-flow: https://danielkummer.github.io/git-flow-cheatsheet/
-
 Please refer to the guidelines_ file if you want more information about process!
 
-.. _guidelines: https://github.com/prise-3d/macop/blob/master/CONTRIBUTING.md 
+.. _guidelines: https://github.com/prise-3d/macop/blob/master/CONTRIBUTING
