+++ docs/source/examples/ubqp/index.rst
@@ -1,0 +1,13 @@
+Unconstrained Binary Quadratic Programming
+==========================================
+
+This example will be available soon.
+
+.. toctree::
+   :maxdepth: 1
+   :numbered:
+   :caption: Contents:
+
+   problem
+   instance
+   implementation
