### 2021-01-26 3c485ed
Add of UBQP instance and unit test for new evaluators
