+++ docs/source/examples/ubqp/index.rst
@@ -1,7 +1,9 @@
 Unconstrained Binary Quadratic Programming
 ==========================================
 
-This example will be available soon.
+Given a collection of :math:`n` items such that each pair of items is associated with a profit value that can be positive, negative or zero, unconstrained binary quadratic programming (UBQP) seeks a subset of items that maximizes the sum of their paired values. The value of a pair is accumulated in the sum only if the two corresponding items are selected.
+
+The UBQP problem will be tackle in this example.
 
 .. toctree::
    :maxdepth: 1
