### 2021-01-26 923a4b8
add date of last release on googlecloud of ceres solver
