+++ paper.bib
@@ -136,6 +136,8 @@
 @misc{ceres-solver,
   author = "Sameer Agarwal and Keir Mierle and Others",
   title = "Ceres Solver",
+  version = "2.0.0",
+  year = "2020",
   howpublished = "\url{http://ceres-solver.org}",
 }
 
