### 2021-01-26 #2812 @jbuisine
Hello @stsievert, @torressa,

I made some updates:

- I updated the paper by naming my thesis supervisors.
- I have added some examples with generated QAP and UBQP instances. They are available in the "Some examples" section of the documentation.
- I have updated the CONTRIBUTING document to make it lighter and more understandable.

Do not hesitate to let me know if I need to change anything about the paper.
