### 2021-01-26 #2812 @stsievert
Thanks for the edits @jbuisine. The documentation is certainly much more descriptive. My comments this time are much smaller; in general, my comments are mostly suggestions. However, I have one blocking comments:

* **Community guidelines** in the recent update, there's no information on users looking to "seek support" (not to my reading at least).
    * Also, CONTRIBUTING not rendered as plain text on GitHub (renaming to CONTRIBUTING.md will fix that).
* **Statement of need:** The audience description of this software is much better – the modularity makes sense. However, the description is still very general.
    * Who would specifically use this software?
    * What are the main features of Macop, and when are they relevant?
    * I think the [Dask docs][1] are a good example, though it doesn't explicitly mention the audience.

I am still not finished with the review; I need to look closer at the implementation. I also have comments on items that technically meet the requirements:

* **Functionality documentation:**
    * Classes don't document initialization methods. Are the arguments stored as class attributes?
    * For example, is it required that [macop.evaluators.discrete.mono][2] take a dictionary? Type annotations would help a lot.
* **Usage:** I have tried to use your software. I stepped through the example at https://jbuisine.github.io/macop/_build/html/documentations/index.html.  This is borderline documentation. Specific comments:
    * My development time is bounded. I'd like a simple self-contained script that I can customize -- I'd rather not dig into the different pages.  (this comment might be outdated with the recent addition of "some examples").
    * Why is necessary to include the pseudo-implementations? (e.g., https://jbuisine.github.io/macop/_build/html/documentations/solutions.html).  Why not link to the docs for macop.solutions.base?

I have some more specific notes/style nits:

* I would further collapse the knapsack example on the docs into a single HTML page. I found it annoying to click to read a short HTML summary for the first couple sections. (note: it is possible to link to headings/sub-headings).
* Why is `solution.{_data, _algo}` private? Shouldn't it be public? They're both in the tests and the documentation, which is an indication they should be public.
* I think the "documentation" link in the documentation sidebar is a meaningless name (of course it's documentation!). Maybe "Example usage"?
* I found it unnecesary to include the class implementations at
    * This page has unnecessary code that the user doesn't need to write: https://jbuisine.github.io/macop/_build/html/documentations/operators.html
    * I think a good example that only writes code the user needs to see is https://jbuisine.github.io/macop/_build/html/documentations/evaluators.html
* It'd be really useful to add type annotations to the docs/examples.
* I think `_build` can be removed from the URL; I think you can specify the build location with Sphinx. I'd recommend changing `https://jbuisine.github.io/macop/_build/html/documentations/index.html` to `https://jbuisine.github.io/macop/html/documentations/index.html` or `https://jbuisine.github.io/macop/documentations/index.html`.
* For `macop.policies.reinforcement.UCBPolicy`, does larger `C`/`exp_rate` mean more or less exploration? What are typical values?
* It'd be really useful it you could provide easier links between types in the documentation (probably easiest by showing more in the table of contents).
    * i.e., the documentation for macop.algorithms.mono.HillClimberBestImprovement says the `evaluator` method takes an object of type `Evaluator`. I can't easily navigate to see that page.
* https://jbuisine.github.io/macop/_build/html/macop/macop.operators.continuous.mutators.html#module-macop.operators.continuous.mutators is empty.
* Docstring for `macop.algorithms.mono`: why are `_score` and `_data` private variables? If they're mentioned in the docs and tested in the docstrings, I think they should be public (without the underscore).
* Paper, line 60: typo with `macop.algorihtms` in header.
* I find comments in the examples like `# policy import` and `# initializer function with lambda function` to be meaningless; that doesn't provide any new information -- `initializer = lambda x=20: ...` tells me the same information.  (I'm looking at https://jbuisine.github.io/macop/_build/html/macop/macop.algorithms.mono.html).  I can't read that example because of the comments; don't docstrings allow newlines to separate the code into "paragraphs"?

[2]:https://jbuisine.github.io/macop/_build/html/macop/macop.evaluators.discrete.mono.html#module-macop.evaluators.discrete.mono
[1]:https://distributed.dask.org
