### 2021-01-27 a0aaa4b
Fix css logo issue
