### 2021-01-27 36c7bae
update of documentation motivation
