+++ macop/evaluators/base.py
@@ -10,7 +10,17 @@ from abc import abstractmethod
 class Evaluator():
     """Abstract Evaluator class which enables to compute solution using specific `_data` 
     """
-    def __init__(self, data):
+    def __init__(self, data: dict):
+                """Apply the computation of fitness from solution
+
+        Fitness is a float value for mono-objective or set of float values if multi-objective evaluation
+
+        Args:
+            solution: {Solution} -- Solution instance
+
+        Return:
+            {float} -- computed solution score (float or set of float if multi-objective evaluation)
+        """
         self._data = data
 
     @abstractmethod
