### 2021-01-27 b0b1450
Update comments and enable use of getter and setter
