+++ docs/source/documentations.rst
@@ -92,7 +92,7 @@ In a first step, the management of the solutions by the macop package will be pr
 Generic Solution
 ~~~~~~~~~~~~~~~~~~~~~~~~~
 
-Inside ``macop.solutions.base`` module of `Macop`, the ``Solution`` class is available. It's an abstract solution class structure which:
+Inside macop.solutions.base_ module of `Macop`, the ``Solution`` class is available. It's an abstract solution class structure which:
 
 - stores the solution data representation into its ``data`` attribute
 - get ``size`` (shape) of specific data representation
@@ -154,7 +154,7 @@ We will now see how to define a type of solution specific to our problem.
 Solution representation for knapsack
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-We will now use the abstract ``Solution`` type available in the ``macop.solutions.base`` module in order to define our own solution.
+We will now use the abstract ``Solution`` type available in the macop.solutions.base_ module in order to define our own solution.
 First of all, let's look at the representation of our knapsack problem. **How to represent the solution?**
 
 Knapsack solution
@@ -177,7 +177,7 @@ where selected objects have **1** as value otherwise **0**.
 Binary Solution
 **********************
 
-We will now define our own type of solution by inheriting from ``macop.solutions.base.Solution``, which we will call ``BinarySolution``.
+We will now define our own type of solution by inheriting from macop.solutions.base.Solution_, which we will call ``BinarySolution``.
 
 First we will define our new class as inheriting functionality from ``Solution`` (such as child class). 
 We will also have to implement the ``random`` method to create a new random solution.
@@ -212,7 +212,7 @@ We will also have to implement the ``random`` method to create a new random solu
             return solution
 
 .. note::
-    The current developed ``BinarySolution`` is available into ``macop.solutions.discrete.BinarySolution`` in **Macop**.
+    The current developed ``BinarySolution`` is available into macop.solutions.discrete.BinarySolution_ in **Macop**.
 
 Using this new Solution representation, we can now generate solution randomly:
 
@@ -299,7 +299,7 @@ Now that it is possible to generate a solution randomly or not. It is important 
 Generic evaluator
 ~~~~~~~~~~~~~~~~~~~~~~
 
-As for the management of solutions, a generic evaluator class ``macop.evaluators.base.Evaluator`` is developed within **Macop**:
+As for the management of solutions, a generic evaluator class macop.evaluators.base.Evaluator_ is developed within **Macop**:
 
 Abstract Evaluator class is used for computing fitness score associated to a solution. To evaluate all the solutions, this class:
 
@@ -385,7 +385,7 @@ It is now possible to initialise our new evaluator with specific data of our pro
     solution_fitness = solution.fitness()
 
 .. note::
-    The current developed ``KnapsackEvaluator`` is available into ``macop.evaluators.mono.KnapsackEvaluator`` in **Macop**.
+    The current developed ``KnapsackEvaluator`` is available into macop.evaluators.mono.KnapsackEvaluator_ in **Macop**.
 
 In the next part we will see how to modify our current solution with the use of modification operator.
 
@@ -402,7 +402,7 @@ In the discrete optimisation literature, we can categorise operators into two se
 - **mutators**: modification of one or more elements of a solution from its current state.
 - **crossovers**: Inspired by Darwin's theory of evolution, we are going here from two solutions to generate a so-called offspring solution composed of the fusion of the data of the parent solutions.
 
[…]
