### 2021-01-27 #2812 @jbuisine
Hello @stsievert,

Thank you very much for your remarks. I updated the package consequently, here my answers on each suggested points:

## Blocking comments

**[RV]**
- Community guidelines in the recent update, there's no information on users looking to "seek support" (not to my reading at least).
    Also, CONTRIBUTING not rendered as plain text on GitHub (renaming to CONTRIBUTING.md will fix that).

**[AA]**
I have added the support seeking information and renamed the contribution file correctly.

**[RV]**
- Statement of need: The audience description of this software is much better – the modularity makes sense. However, the description is still very general.
    Who would specifically use this software?
    What are the main features of Macop, and when are they relevant?
    I think the [Dask docs](https://distributed.dask.org/en/latest/) are a good example, though it doesn't explicitly mention the audience.

**[AA]**

I add the motivation of the package in [description](https://jbuisine.github.io/macop/_build/html/) of the documentation. I was indeed inspired by the approach proposed for the Dask software. I also add this part into the paper file.

**[RV]**
- Functionality documentation:
    Classes don't document initialization methods. Are the arguments stored as class attributes?
    For example, is it required that macop.evaluators.discrete.mono take a dictionary? Type annotations would help a lot.

**[AA]**

I have updated the documentation by completing the comments on instantiation methods. In addition, I have added typing the `data` parameter of the Evaluator class as a `dict` as advised.

**[RV]**
- Usage: I have tried to use your software. I stepped through the example at https://jbuisine.github.io/macop/_build/html/documentations/index.html. This is borderline documentation. Specific comments:
    My development time is bounded. I'd like a simple self-contained script that I can customize -- I'd rather not dig into the different pages. (this comment might be outdated with the recent addition of "some examples").
    Why is necessary to include the pseudo-implementations? (e.g., https://jbuisine.github.io/macop/_build/html/documentations/solutions.html). Why not link to the docs for macop.solutions.base?

**[AA]**

Indeed, the proposed tutorial allows a complete tour of the use of the Macop's interaction loop. For that, I made it quite complete. As you specified, now there are many faster examples on adding and processing specific [problem instances](https://jbuisine.github.io/macop/_build/html/examples.html). They are therefore faster to implement and require less time.

In documentation, API links to module have also been added.

## Specific notes/style nits:

**[RV]**

  - I would further collapse the knapsack example on the docs into a single HTML page. I found it annoying to click to read a short HTML summary for the first couple sections. (note: it is possible to link to headings/sub-headings).

**[AA]**

I updated the documentation consequently and keep the whole "tutorial" in the single with complete menu in left sidebar.

**[RV]**

  - Why is `solution.{_data, _algo}` private? Shouldn't it be public? They're both in the tests and the documentation, which is an indication they should be public.

**[AA]**

I have modified the scope of the variables if needed or added accessors/setters if needed. I think it will indeed be more logical to have the code partitioned in the current way.

**[RV]**

  - I think the "documentation" link in the documentation sidebar is a meaningless name (of course it's documentation!). Maybe "Example usage"?

**[AA]**

Indeed, I updated the title into "Tour of Macop", as we take a complete tour of how the package works.

**[RV]**

  - I found it unnecesary to include the class implementations at
        This page has unnecessary code that the user doesn't need to write: [https://jbuisine.github.io/macop/_build/html/documentations/operators.html](https://jbuisine.github.io/macop/_build/html/documentations/operators.html)
        I think a good example that only writes code the user needs to see is [https://jbuisine.github.io/macop/_build/html/documentations/evaluators.html](https://jbuisine.github.io/macop/_build/html/documentations/evaluators.html)

**[AA]**

I know this is a bit redundant, but it was in the idea of allowing the user following this tutorial to have access to the code and the evolution of the one with the knapsack problem.

**[RV]**

  - It'd be really useful to add type annotations to the docs/examples.

**[AA]**

I don't really understand what you want to say. Type annotations depending of solutions obtained?

**[RV]**

  - I think `_build` can be removed from the URL; I think you can specify the build location with Sphinx. I'd recommend changing `https://jbuisine.github.io/macop/_build/html/documentations/index.html` to `https://jbuisine.github.io/macop/html/documentations/index.html` or `https://jbuisine.github.io/macop/documentations/index.html`.

**[AA]**

I confess that I haven't found a way to modify this part either on the side of Github or Sphinx. Sphinx allows in its Makefile to specify the output folder but that seems to be all (hence the `_build` output).

**[RV]**
  - For `macop.policies.reinforcement.UCBPolicy`, does larger `C`/`exp_rate` mean more or less exploration? What are typical values?

**[AA]**

Comments `macop.policies.reinforcement.UCBPolicy` have been added in order to explain the interest of these parameters with example values and their impacts.

**[RV]**
  - It'd be really useful it you could provide easier links between types in the documentation (probably easiest by showing more in the table of contents).
        i.e., the documentation for `macop.algorithms.mono.HillClimberBestImprovement` says the evaluator method takes an object of type `Evaluator`. I can't easily navigate to see that page.

**[AA]**

In documentation, API links to module have also been added in order to give a better navigration.

**[RV]**
  - [https://jbuisine.github.io/macop/_build/html/macop/macop.operators.continuous.mutators.html#module-macop.operators.continuous.mutators](https://jbuisine.github.io/macop/_build/html/macop/macop.operators.continuous.mutators.html#module-macop.operators.continuous.mutators) is empty.

**[AA]**

This is intended to make it easier for users already wishing to implement continuous optimization problems.

**[RV]**
  - Docstring for `macop.algorithms.mono`: why are `_score` and `_data` private variables? If they're mentioned in the docs and tested in the docstrings, I think they should be public (without the underscore).

**[AA]**

I have modified the scope of the variables if needed or added accessors/setters if needed. I think it will indeed be more logical to have the code partitioned in the current way.

**[RV]**
  - Paper, line 60: typo with `macop.algorihtms` in header.

**[AA]**

Typo issue fixed into paper.

**[RV]**:

  - I find comments in the examples like `# policy import` and `# initializer function with lambda` function to be meaningless; that doesn't provide any new information -- `initializer = lambda x=20: ...` tells me the same information. (I'm looking at [https://jbuisine.github.io/macop/_build/html/macop/macop.algorithms.mono.html](https://jbuisine.github.io/macop/_build/html/macop/macop.algorithms.mono.html)). I can't - - read that example because of the comments; don't docstrings allow newlines to separate the code into "paragraphs"?

**[AA]**

I have updated some of the comments and get a better visibility for doctest examples.

---

Do not hesitate to let me know if some of my answers are not correct. Thanks.
