### 2021-01-28 #2812 @stsievert
My replies are below. The documentation is looking a lot better.

> I add the motivation of the package in description of the documentation. I was indeed inspired by the approach proposed for the Dask software. I also add this part into the paper file.

Thank you. It's a lot better now. It'd be nice to have a specific example of a use case (but that's not critical).

> the documentation for macop.algorithms.mono.HillClimberBestImprovement says the evaluator method takes an object of type Evaluator. I can't easily navigate to see that page.
> ...

One example: I mean that I can't easily navigate from [macop.algorithms.mono.HillClimberFirstImprovment][1] to [macop.evaluators.base][2], even thought it's a returned by `HillClimberFirstImprovment.evaluator`.
I think this is possible with <code>ref:\`~macop.evaluators.base.Evaluator\`</code> in Sphinx.

----

>  Why is solution.{_data, _algo} private? Shouldn't it be public? They're both in the tests and the documentation, which is an indication they should be public.

I've seen your solution. This code is a little cleaner:

``` python
class Algorithm:
    @property
    def data(self):
        return self._data
```

> It'd be really useful to add type annotations to the docs/examples. ... I don't really understand what you want to say. Type annotations depending of solutions obtained?

For example, `macop.evaluators.base.Evaluator.compute` takes an argument of type `Solution`. Presumably that's an instance of `macop.solutions.base.Solution`. It'd be helpful to have a type hint on that (and even more useful if it's an HTML link).

* **New comment:** it'd be great if you could link to `macop.policies.reinforcement.UCBPolicy` instead of italicizing that text.

> https://jbuisine.github.io/macop/_build/html/macop/macop.operators.continuous.mutators.html is empty. ... This is intended to make it easier for users already wishing to implement continuous optimization problems.

How does that make it easier for users? To modify Macop, they have to write code/tests and submit a PR, or do a development install.

[3]:https://github.com/jbuisine/macop/blob/dff57110ee32870bf52632261abff33441df30e7/docs/Makefile#L8
[1]:https://jbuisine.github.io/macop/_build/html/macop/macop.algorithms.mono.html#macop.algorithms.mono.HillClimberFirstImprovment
[2]:https://jbuisine.github.io/macop/_build/html/macop/macop.evaluators.base.html#module-macop.evaluators.base
