### 2021-01-28 9e997ac
Update of Solution and Algorithm attributes using @property decorator
