+++ macop/algorithms/multi.py
@@ -22,16 +22,16 @@ class MOEAD(Algorithm):
         nObjectives: {int} -- number of objectives (based of number evaluator)
         initialiser: {function} -- basic function strategy to initialise solution
         evaluator: {[function]} -- list of basic function in order to obtained fitness (multiple objectives)
-        operators: {[Operator]} -- list of operator to use when launching algorithm
-        policy: {Policy} -- Policy class implementation strategy to select operators
+        operators: {[:class:`~macop.operators.base.Operator`]} -- list of operator to use when launching algorithm
+        policy: {:class:`~macop.policies.base.Policy`} -- Policy class implementation strategy to select operators
         validator: {function} -- basic function to check if solution is valid or not under some constraints
         maximise: {bool} -- specify kind of optimisation problem
         verbose: {bool} -- verbose or not information about the algorithm
-        population: [{Solution}] -- population of solution, one for each sub problem
-        pfPop: [{Solution}] -- pareto front population
+        population: [{:class:`~macop.solutions.base.Solution`}] -- population of solution, one for each sub problem
+        pfPop: [{:class:`~macop.solutions.base.Solution`}] -- pareto front population
         weights: [[{float}]] -- random weights used for custom mu sub problems
-        callbacks: {[Callback]} -- list of Callback class implementation to do some instructions every number of evaluations and `load` when initialising algorithm
-        parent: {Algorithm} -- parent algorithm reference in case of inner Algorithm instance (optional)
+        callbacks: {[:class:`~macop.callbacks.base.Callback`]} -- list of Callback class implementation to do some instructions every number of evaluations and `load` when initialising algorithm
+        parent: {:class:`~macop.algorithms.base.Algorithm`} -- parent algorithm reference in case of inner Algorithm instance (optional)
 
     >>> import random
     >>>
@@ -58,7 +58,7 @@ class MOEAD(Algorithm):
     >>>
     >>> # validator specification (based on weights of each objects)
     >>> weights = [ random.randint(5, 30) for i in range(problem_size) ]
-    >>> validator = lambda solution: True if sum([weights[i] for i, value in enumerate(solution.getData()) if value == 1]) < 200 else False
+    >>> validator = lambda solution: True if sum([weights[i] for i, value in enumerate(solution.data) if value == 1]) < 200 else False
     >>>
     >>> # initialiser function for binary solution using specific solution size
     >>> initialiser = lambda x=20: BinarySolution.random(x, validator)
@@ -95,11 +95,11 @@ class MOEAD(Algorithm):
             T: {[float]} -- number of neightbors for each sub problem
             initialiser: {function} -- basic function strategy to initialise solution
             evaluator: {[function]} -- list of basic function in order to obtained fitness (multiple objectives)
-            operators: {[Operator]} -- list of operator to use when launching algorithm
-            policy: {Policy} -- Policy class implementation strategy to select operators
+            operators: {[:class:`~macop.operators.base.Operator`]} -- list of operator to use when launching algorithm
+            policy: {:class:`~macop.policies.base.Policy`} -- Policy class implementation strategy to select operators
             validator: {function} -- basic function to check if solution is valid or not under some constraints
             maximise: {bool} -- specify kind of optimisation problem
-            parent: {Algorithm} -- parent algorithm reference in case of inner Algorithm instance (optional)
+            parent: {:class:`~macop.algorithms.base.Algorithm`} -- parent algorithm reference in case of inner Algorithm instance (optional)
             verbose: {bool} -- verbose or not information about the algorithm
         """
 
@@ -215,7 +215,7 @@ class MOEAD(Algorithm):
             evaluations: {int} -- number of Local search evaluations
             
         Returns:
-            {Solution} -- best solution found
[…]
