### 2021-01-28 #2812 @jbuisine
Hi @stsievert ,

Thanks for your quick answer and advices. I have made some updates:

---

> Thank you. It's a lot better now. It'd be nice to have a specific example of a use case (but that's not critical).

Inside README.md file, I include a quick explanation why Macop is used inside an other project. It has also been added inside the paper.md file.

> One example: I mean that I can't easily navigate from macop.algorithms.mono.HillClimberFirstImprovment to macop.evaluators.base, even thought it's a returned by HillClimberFirstImprovment.evaluator.

I use `` :class:`~macop.XXXX.XXXX` `` inside API documentation. It is true that this makes browsing more pleasant. Thanks for the advice.

> I've seen your solution. This code is a little cleaner:

I use `@property` for specific attributes:
- Inside `Soution`: `data` function (set as `@property` and `@data.setter`).
- Inside `Soution`: `fitness` function (set as `@property` and `@fitness.setter`).
- Inside `Algorithm`: `result` function (set as `@property` and `@result.setter`).

The aim is to make it possible to return and process the data/result that the user wants if he creates a new solution/new algorithm.

I have also updated the "A tour of Macop" documentation regarding these changes.

> **New comment:** it'd be great if you could link to macop.policies.reinforcement.UCBPolicy instead of italicizing that text.

This has been updated.

> How does that make it easier for users? To modify Macop, they have to write code/tests and submit a PR, or do a development install.

Yes, the user should be able to propose his modifications if he wants them to be integrated in a new version. The idea was to already fix the way the package would be structured to accommodate such modifications.
