### 2021-01-28 #2812 @stsievert
Nit: https://jbuisine.github.io/macop/_build/html/macop/macop.policies.classicals.html does not completely describe the API for `RandomPolicy`: it's missing the `setAlgo` and `apply` methods.

It looks like this package has two packs: the discrete algorithms/operators/etc, and the data flow that allows customization. Maybe it'd be better to have that separation on the docs? i.e., have all the base classes and other classes well separated in separate sections/page.

> Yes, the user should be able to propose his modifications if he wants them to be integrated in a new version. The idea was to already fix the way the package would be structured to accommodate such modifications.

I'm looking at this package from a user perspective, not a developer or contributor perspective. I think an empty module (e.g., `macop.solutions.continuous`) only confuses users (like me) when they click on the documentation link. I think contributors will contact you with a beta implementation, and you can guide them (either in CONTRIBUTING.md or in the pull request).
