+++ .github/workflows/docs.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:    
       - master
-      - develop
 
 jobs:
   build:
