### 2021-01-29 #2812 @jbuisine
Thans for your answers.

Here my relative updates:

- Display of inheritance classes and members inside API for each class (using `'inherited-members', 'show-inheritance'` keywords for autosummary)

- API documentation page is now divided in 4 sections:
   - Modules description
   - Common and base modules
   - Discrete Optimisation
   - Continuous Optimisation (with some explanations why and guidelines link)
