### 2021-02-01 7309892
Update UCB policy default exp value
