### 2021-02-03 #2812 @jbuisine
Hi @stsievert, @torressa,

I wanted to inform you that I have also updated the paper (only a few changes).
