### 2021-02-20 #2812 @torressa
Last thing, the badges in the readme file should lead to appropriate links (as the JOSS badge does).
Other than that I'm happy to recommend acceptance :)
