### 2021-02-22 #2812 @stsievert
I have followed "[A Tour of Macop][1]" and tried to implement a continuous optimization problem (gradient descent on a restricted search space). I got close to making it work, but not quite. I have some specific comments, both of which are blocking:

1. **Functionality documentation:**  the functionality is not documented. I tried walking through the example to test one of the core features, the generalization to continuous problems (mentioned explicitly in the paper and docs). I couldn't get this working.
    * (I'd recommend providing another documentation page to "A Tour of Macop," but keeping it very simple (then linking to "A Tour of Macop" for more advanced usage). If generalizing to continuous optimization is explicitly mentioned, I'd walk through an instance of that problem).
2. **A statement of need:**  The authors do clearly state the problem the software is designed to solve. However, they do not indicate the target audience.
    * (I think this could be resolved by answering "who cares about flexibility in discrete optimization?").

Additionally, I have some specific style nits (all of which are from "A Tour of Macop"):

* In "A Tour of Macop", the base classes (e.g., `class Solution: ...`) is distracting/confusing and doesn't provide much value.
* I wouldn't recommend mentioning a private attribute on the docs (e.g., `_data` attribute of evaluators). I'd make these attributes public attributes; mentioning them on the docs is a good indicator they should be public.
* Evaluators: use `score` instead of `compute`? Of course `compute` performs computation. What does it compute?
* `self` not defined in KnapsackEvaluator.
* I found the code at "local search algorithm" to be really useful. I would link to that section from the top of the docs.

[1]:https://jbuisine.github.io/macop/_build/html/documentations.html
