+++ examples/ZdtExample.py
@@ -3,19 +3,20 @@ import logging
 import os
 import random
 import numpy as np
+import math
 
 # module imports
-from macop.solutions.discrete import BinarySolution
-from macop.evaluators.discrete.mono import UBQPEvaluator
+from macop.solutions.continuous import ContinuousSolution
+from macop.evaluators.continuous.mono import ZdtEvaluator
 
-from macop.operators.continuous.mutators import SimpleMutation
-from macop.operators.discrete.mutators import SimpleBinaryMutation
+from macop.operators.continuous.mutators import PolynomialMutation
+from macop.operators.continuous.crossovers import BasicDifferentialEvolutionCrossover
 
 from macop.policies.classicals import RandomPolicy
 
 from macop.algorithms.mono import IteratedLocalSearch as ILS
 from macop.algorithms.mono import HillClimberFirstImprovment
-from macop.callbacks.classicals import BasicCheckpoint
+from macop.callbacks.classicals import ContinuousCheckpoint
 
 if not os.path.exists('data'):
     os.makedirs('data')
@@ -26,52 +27,47 @@ logging.basicConfig(format='%(asctime)s %(message)s', filename='data/example.log
 random.seed(42)
 
 # usefull instance data
-n = 100
-ubqp_instance_file = 'instances/ubqp/ubqp_instance.txt'
-filepath = "data/checkpoints_ubqp.csv"
+n = 10
+filepath = "data/checkpoints_zdt_Rosenbrock.csv"
+problem_interval = -10, 10 # fixed value interval (avoid infinite)
 
 
-# default validator
+# check each value in order to validate
 def validator(solution):
+
+    mini, maxi = problem_interval
+
+    for x in solution.data:
+        if x < mini or x > maxi:
+            return False
+
     return True
 
 # define init random solution
 def init():
-    return BinarySolution.random(n, validator)
-
-
-filepath = "data/checkpoints.csv"
+    return ContinuousSolution.random(n, problem_interval, validator)
 
 def main():
 
-    # load UBQP instance
-    with open(ubqp_instance_file, 'r') as f:
+    # Rosenbrock function with a=1 and b=100 (see https://en.wikipedia.org/wiki/Rosenbrock_function)
+    Rosenbrock_function = lambda s: sum([ 100 * math.pow(s.data[i + 1] - (math.pow(s.data[i], 2)), 2) + math.pow((1 - s.data[i]), 2) for i in range(len(s.data) - 1) ])
 
-        lines = f.readlines()
-
-        # get all string floating point values of matrix
-        Q_data = ''.join([ line.replace('\n', '') for line in lines[8:] ])
-
-        # load the concatenate obtained string
-        Q_matrix = np.fromstring(Q_data, dtype=float, sep=' ').reshape(n, n)
-
-    print(f'Q_matrix shape: {Q_matrix.shape}')
-
-    operators = [SimpleBinaryMutation(), SimpleMutation()]
+    operators = [PolynomialMutation(interval=problem_interval), BasicDifferentialEvolutionCrossover(interval=problem_interval)]
     policy = RandomPolicy(operators)
-    callback = BasicCheckpoint(every=5, filepath=filepath)
-    evaluator = UBQPEvaluator(data={'Q': Q_matrix})
+    callback = ContinuousCheckpoint(every=5, filepath=filepath)
+    evaluator = ZdtEvaluator(data={'f': Rosenbrock_function})
 
     # passing global evaluation param from ILS
-    hcfi = HillClimberFirstImprovment(init, evaluator, operators, policy, validator, maximise=True, verbose=True)
-    algo = ILS(init, evaluator, operators, policy, validator, localSearch=hcfi, maximise=True, verbose=True)
[…]
