+++ macop/__init__.py
@@ -9,6 +9,8 @@ from macop.callbacks import multi
 from macop.evaluators import base
 from macop.evaluators.discrete import mono
 from macop.evaluators.discrete import multi
+from macop.evaluators.continuous import mono
+from macop.evaluators.continuous import multi
 
 from macop.operators import base
 from macop.operators.discrete import mutators
