### 2021-02-24 afadd95
update some private attributes as public
