+++ paper.md
@@ -53,6 +53,17 @@ Hence, motivation behind **Macop** is a flexible discrete optimisation package a
 - **Extensible:** the package is open to extension, i.e. it does not partition the user in these developer choices. It can just as well implement continuous optimization problems if needed while making use of the main interaction loop proposed by the package.
 - **Easy Setup:** as a pure Python package distributed is `pip` installable and easy to use.
 
+# Target Audience 
+
+This package would meet the expectations of people wishing to: 
+- Solve a complex problem oriented evolutionary algorithm but who do not wish to develop their own framework. They can rely on what the package already proposes but also on its generic and flexible contribution in order to adapt their own content;
+- Conduct research work leading to the rapid modification of meta-heuristics and the interaction of different algorithms. More precisely:
+  - test new combinations of algorithms. Changing algorithms during evaluations, e.g. different local searches;
+  - provide reinforcement learning during searches (e.g. adaptive operator choice strategy).
+  - test new multi-objective methods quickly thanks to the proposed algorithmic hierarchy allowing to easily decompose the multi-objective problem into single-objective sub-problems.
+- Take advantage of a system for launching calculations from a backup in order to avoid any loss in case of unwanted program interruption;
+- Quickly model a problem that is still unknown, i.e. the type of solution and the evaluation function, while taking advantage of the interaction loop proposed by the package.
+
 # Description
 
 At the beginning of the development of this library, the idea of making it as modular as possible was topical. The library divide into sub-module forms considered to be the most important to build and solve an optimisation problem.
