### 2021-02-24 #2812 @jbuisine
Thanks @stsievert for your feedback.

I made some updates. Concerning:

1.  **Functionnality documentation:**
Examples of the implementation of continuous optimisation problems were proposed. With an example of a Zdt function to be optimised. The indications are available at the end of "A Tour of Macop", in the "Advanced usages" section which includes discrete multi-objective implementations and the continuous optimisation example.
2. **A statement of need:**
I have taken up your question: "who cares about flexibility in discrete optimization" and have provided an answer in the "Target Audience" section of the README.md, the paper and the documentation.

In "A Tour of Macop":
- the base classes (e.g., class Solution: ...) have been removed. Indeed, the links to the documentation are already sufficient. Only the more specific codes are present;
- I changed some attributes as public inside the package such as ``_data``, ``_algo`` and ``_operators``...;
- I fixed the problem for Evaluator explanation (computation of fitness and self attribute of `compute` method).
