### 2021-02-24 #2812 @stsievert
Thanks for the quick response @jbuisine. It appears the changes are addressed.

1. **Documentation:** Thanks for the examples in "advanced usage." I think resolves the issue I had. As a new user of Macop, I would appreciate if that material were more prominent (maybe a new doc page titled "example uses"?)
2. **A statement of need/target audience:** The provided description is very complete and very detailed; to be frank, more than I thought you'd write. I expected something similar to "Macop would provide use in data scientists looking to solve delivery vehicle routing problems, or for economists looking to optimize their automated trading system" (though I'm not to sure on the specific applications).

I'll try to finish my review this weekend (which will require using the docs to implement a new example).
