### 2021-02-24 #2812 @jbuisine
@stsievert Thanks for your quick response too.

Concerning the documentation I renamed the section "Some examples" into "Example uses" add I added a full page for the [Zdt continuous Rosenbrock problem instance](https://jbuisine.github.io/macop/_build/html/examples.html). I hope that this is better for a visual of the available examples.
