### 2021-03-02 #2812 @stsievert
I've checked off all boxes for review; I now recommend this software for acceptance.

I do not think v1.0.0 of this software package should be accepted. Can we change the version to v1.1.0 or v1.0.16?
