### 2021-03-02 #2812 @melissawm
Yes; if @jbuisine does a tagged release and reports the updated version number here, that's the version that is going to be archived along with the JOSS paper.
