### 2021-03-03 f757238
Using __init__ to set version of documentation too
