+++ docs/source/conf.py
@@ -15,9 +15,9 @@
 import os
 import sys
 import asteroid_sphinx_theme
-import macop
 
 sys.path.insert(0, os.path.abspath('../../../macop'))
+import macop
 
 # -- Project information -----------------------------------------------------
 
