### 2021-03-03 73541b2
Import only __version__ when using tag
