+++ macop/solutions/__init__.py
@@ -1,0 +1,3 @@
+from macop.solutions import base
+from macop.solutions import continuous
+from macop.solutions import discrete
