+++ macop/utils/__init__.py
@@ -1,0 +1,1 @@
+from macop.utils import progress
