+++ macop/callbacks/__init__.py
@@ -1,0 +1,3 @@
+from macop.callbacks import base
+from macop.callbacks import classicals
+from macop.callbacks import multi
