### 2021-03-03 3332a9f
Resize Macop logo
