### 2021-03-03 #2812 @jbuisine
Thank you all for all your comments and the time you have given me which has enabled me to improve the package.

I have updated the latest version of the package v1.2.0. I had a small problem with version 1.1.0.

Have a nice day,
