### 2021-03-05 #2812 @melissawm
Thank you @jbuisine - there are a few final corrections to the paper.

You do need a section named `Statement of Need` (this is mentioned in the [JOSS Submission Guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). My suggestion is to rename your Motivation section to that.

There are a few remaining typos, some of them may be better verified with a spellchecker. I ask that you do a careful review of those. The following is a list of suggestions.

- L12 There is a small typo: `In this paper, Macop for Minimalist And Customisable Optimisation Package, is a` (the final `a` should be removed)
- L17 I believe `Based on a common interaction loop (see Figure 1) of all the algorithms. Macop wants to` should read `Based on a common interaction loop (see Figure 1) of all the algorithms, Macop wants to`
- L58-59 I am not sure what this first sentence means, can you rephrase or clarify?
- L80-81 I believe this should read `validator: such as constraint programming, a validator is a function which is used to validate or not a solution data state;`
- L82-83 Should read `evaluator: stores problem instance data and implements a compute method in order to evaluate a solution;`
- L84 I believe this should read `operators: mutators, crossover operators to update and obtain a new solution`
- L100-103 I believe this sentence is incomplete, perhaps it can be rephrased to be clearer.
- L104 I think this should read `to change multi-objectives problem into a set of mono-objective problems`
- L106 Is *mu* supposed to be rendered as text, or as a symbol?
- L115 Should read `anyone who wants to model his own problem.`
- L117-118 I think this should read `A few mutation and crossover operators have been implemented. However, it remains quite simple.`
- L124 Should read `The use of callback instance allows both to do an action every k evaluations of information`
- L127-128 I believe this should read `As an example, the implemented UCBPolicy has its own callback...`
- L130-132 It is not necessary to link back to the package documentation. The link to the package repository is contained in the DOI and paper metadata.
- L134 Should read `Macop aims to allow the modelling of discrete (usually combinatorial) optimisation problems.`

In the References section, please review the following:
- In the reference (Alves, M. J., & Almeida, M. (2007)), Thebycheff should be capitalized.
- Please verify the citation for Pyomo - it should read `Second Edition`.
- L159 One of the authors' name (Fialho, A.) should be corrected.
- Please verify that the citation on lines 162-163 is in accord with [the publisher](https://link.springer.com/chapter/10.1007%2F0-306-48056-5_11#citeas).
- Please verify that the citation on lines 165-168 is in accord with [the publisher](https://link.springer.com/chapter/10.1007%2F978-3-319-42432-3_37#citeas)
- Can you review both citations for GitHub repositories so they agree with the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) for JOSS?
- Finally, some references contain a full journal name and others contain abbreviated journal names. For consistency, would you mind writing all journal names in the same format?

Thank you again, and let me know if there's anything I can help you with.
