+++ paper.bib
@@ -1,28 +1,25 @@
-@incollection{DBLP:books/sp/03/LourencoMS03,
-  author    = {Helena R. Louren{\c{c}}o and
-               Olivier C. Martin and
-               Thomas St{\"{u}}tzle},
-  editor    = {Fred W. Glover and
-               Gary A. Kochenberger},
-  title     = {Iterated Local Search},
-  booktitle = {Handbook of Metaheuristics},
-  series    = {International Series in Operations Research {\&} Management Science},
-  volume    = {57},
-  pages     = {320--353},
-  publisher = {Kluwer / Springer},
-  year      = {2003},
-  url       = {https://doi.org/10.1007/0-306-48056-5\_11},
-  doi       = {10.1007/0-306-48056-5\_11},
-  timestamp = {Mon, 15 Jun 2020 16:48:23 +0200},
-  biburl    = {https://dblp.org/rec/books/sp/03/LourencoMS03.bib},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
+@Inbook{Lourenço2003,
+  author="Louren{\c{c}}o, Helena R.
+  and Martin, Olivier C.
+  and St{\"u}tzle, Thomas",
+  editor="Glover, Fred
+  and Kochenberger, Gary A.",
+  title="Iterated Local Search",
+  bookTitle="Handbook of Metaheuristics",
+  year="2003",
+  publisher="Springer US",
+  address="Boston, MA",
+  pages="320--353",
+  isbn="978-0-306-48056-0",
+  doi="10.1007/0-306-48056-5_11",
+  url="https://doi.org/10.1007/0-306-48056-5_11"
 }
 
 @article{DBLP:journals/tec/ZhangL07,
   author    = {Qingfu Zhang and
                Hui Li},
   title     = {{MOEA/D:} {A} Multiobjective Evolutionary Algorithm Based on Decomposition},
-  journal   = {{IEEE} Trans. Evol. Comput.},
+  journal   = {{IEEE} Transactions on Evolutionary Computation},
   volume    = {11},
   number    = {6},
   pages     = {712--731},
@@ -38,9 +35,9 @@
 @article{DBLP:journals/cor/AlvesA07,
   author    = {Maria Jo{\~{a}}o Alves and
                Marla Almeida},
-  title     = {{MOTGA:} {A} multiobjective Tchebycheff based genetic algorithm for
+  title     = {{MOTGA:} {A} multiobjective {T}chebycheff based genetic algorithm for
                the multidimensional knapsack problem},
-  journal   = {Comput. Oper. Res.},
+  journal   = {Computers & Operations Research},
   volume    = {34},
   number    = {11},
   pages     = {3458--3470},
@@ -55,12 +52,12 @@
 
 @article{DBLP:journals/tec/LiFKZ14,
   author    = {Ke Li and
-               {\'{A}}lvaro Fialho and
+               \'{A}lvaro Fialho and
                Sam Kwong and
                Qingfu Zhang},
   title     = {Adaptive Operator Selection With Bandits for a Multiobjective Evolutionary
                Algorithm Based on Decomposition},
-  journal   = {{IEEE} Trans. Evol. Comput.},
+  journal   = {{IEEE} Transactions on Evolutionary Computation},
   volume    = {18},
   number    = {1},
   pages     = {114--130},
@@ -121,7 +118,7 @@
   title     = {Integrating Airborne Hyperspectral, Topographic, and Soil Data for
                Estimating Pasture Quality Using Recursive Feature Elimination with
                Random Forest Regression},
-  journal   = {Remote. Sens.},
+  journal   = {Remote Sensing},
   volume    = {10},
   number    = {7},
[…]
