### 2021-03-11 #2812 @melissawm
Thank you! Just one last check: I think the license for the Zenodo archive should be the same as what you use on your code - the MIT License. Can you check the Zenodo archive for that? Otherwise, the metadata looks good.
