### 2021-03-11 #2812 @jbuisine
Indeed I forgot to look at the license. I just modified the Zenodo publication accordingly with MIT Licence.
