### 2021-03-11 #2812 @kyleniemeyer
No worries @melissawm!

Hi @jbuisine, I'm doing some final checks before publishing—can you add that missing DOI that the bot pointed out above?
