+++ paper.bib
@@ -161,15 +161,27 @@
   doi = {10.1007/s00158-011-0666-3}
 }
 
-@incollection{MaherMiltenbergerPedrosoRehfeldtSchwarzSerrano2016,
-  author = {Stephen Maher and Matthias Miltenberger and Jo{\~{a}}o Pedro Pedroso and Daniel Rehfeldt and Robert Schwarz and Felipe Serrano},
-  title = {{PySCIPOpt}: Mathematical Programming in Python with the {SCIP} Optimization Suite},
-  booktitle = {Mathematical Software {\textendash} {ICMS} 2016},
-  publisher = {Springer International Publishing},
-  pages = {301--307},
-  year = {2016},
-  doi = {10.1007/978-3-319-42432-3_37},
+@InProceedings{10.1007/978-3-319-42432-3_37,
+  author="Maher, Stephen
+  and Miltenberger, Matthias
+  and Pedroso, Jo{\~a}o Pedro
+  and Rehfeldt, Daniel
+  and Schwarz, Robert
+  and Serrano, Felipe",
+  editor="Greuel, Gert-Martin
+  and Koch, Thorsten
+  and Paule, Peter
+  and Sommese, Andrew",
+  title="PySCIPOpt: Mathematical Programming in Python with the SCIP Optimization Suite",
+  booktitle="Mathematical Software -- ICMS 2016",
+  year="2016",
+  publisher="Springer International Publishing",
+  address="Cham",
+  pages="301--307",
+  abstract="SCIP is a solver for a wide variety of mathematical optimization problems. It is written in C and extendable due to its plug-in based design. However, dealing with all C specifics when extending SCIP can be detrimental to development and testing of new ideas. This paper attempts to provide a remedy by introducing PySCIPOpt, a Python interface to SCIP that enables users to write new SCIP code entirely in Python. We demonstrate how to intuitively model mixed-integer linear and quadratic optimization problems and moreover provide examples on how new Python plug-ins can be added to SCIP.",
+  isbn="978-3-319-42432-3"
 }
+
 
 @misc{simanneal-solver,
   author = "Matthew Perry",
