### 2021-03-12 #2812 @jbuisine
@melissawm, @kyleniemeyer Hi!

I'm sorry but I do not understand why this reference is not checked. The DOI refers to this [article](https://link.springer.com/chapter/10.1007%2F978-3-319-42432-3_37).

Could you please tell me what I might have forgotten?
