+++ paper.bib
@@ -176,7 +176,8 @@
   address="Cham",
   pages="301--307",
   abstract="SCIP is a solver for a wide variety of mathematical optimization problems. It is written in C and extendable due to its plug-in based design. However, dealing with all C specifics when extending SCIP can be detrimental to development and testing of new ideas. This paper attempts to provide a remedy by introducing PySCIPOpt, a Python interface to SCIP that enables users to write new SCIP code entirely in Python. We demonstrate how to intuitively model mixed-integer linear and quadratic optimization problems and moreover provide examples on how new Python plug-ins can be added to SCIP.",
-  isbn="978-3-319-42432-3"
+  isbn="978-3-319-42432-3",
+  doi="10.1007/978-3-319-42432-3_37",
 }
 
 
